refactor: decompose remaining message handlers from WebSocketMessageHandler
- Extracted the remaining 19 domain handlers into dedicated module files.
- Reduced WebSocketMessageHandler.ts from 6,000+ down to 680 lines, leaving only core routing logic.
- Refactored handlers to take an explicit MessageHandlerContext instead of `this`, with zero behavioral changes.
- Validated with tsc, ESLint, Prettier, full test, and live end-to-end testing on the browser, desktop and CLI.
