+++ src/main/web-server/handlers/messageHandlers/WebSocketMessageHandler.ts
@@ -29,106 +29,10 @@
  * are handled by ./autoRun.
  */
 
-import path from 'path';
-import fs from 'fs/promises';
 import { app } from 'electron';
 import { logger } from '../../../utils/logger';
 import { captureException } from '../../../utils/sentry';
 import { getCommitHash } from '../../../utils/build-info';
-import {
-	startProfiling,
-	stopProfiling,
-	getProfilingStatus,
-	finalizeCapture,
-} from '../../../profiling';
-import { getStatsDB } from '../../../stats/singleton';
-import { runReadonlyStatsQuery } from '../../../stats/readonly-query';
-import type { StatsTimeRange } from '../../../../shared/stats-types';
-import type {
-	SettingValue,
-	CreateSessionConfig,
-	WebPlaybookDocument,
-	NotifyToastClickAction,
-	NotifyToastKind,
-	NotifyToastColor,
-	NotifyCenterFlashColor,
-	NotifyCenterFlashVariant,
-} from '../../types';
-
-/** Canonical Toast / Center Flash color set (shared design language). */
-const NOTIFY_COLORS: readonly NotifyCenterFlashColor[] = [
-	'green',
-	'yellow',
-	'orange',
-	'red',
-	'theme',
-];
-const NOTIFY_FLASH_COLORS = NOTIFY_COLORS;
-const NOTIFY_TOAST_COLORS = NOTIFY_COLORS;
-
-const NOTIFY_TOAST_KINDS: readonly NotifyToastKind[] = ['success', 'info', 'warning', 'error'];
-
-/**
- * Legacy variant/type → color mapping. Lets older CLI scripts keep working
- * while we transition external integrations to `--color`.
- */
-const VARIANT_TO_COLOR: Record<NotifyCenterFlashVariant, NotifyCenterFlashColor> = {
-	success: 'green',
-	info: 'theme',
-	warning: 'yellow',
-	error: 'red',
-};
-
-/**
- * Hard upper bound on flash duration for **externally-triggered** flashes
- * (CLI / web). The renderer-side `notifyCenterFlash` itself is uncapped so
- * internal in-app callers can still use longer durations if ever needed -
- * the cap lives at the IPC boundary so external scripts can't stick a
- * permanent overlay on the user.
- */
-const EXTERNAL_FLASH_MAX_DURATION_MS = 5000;
-
-/**
- * Hard upper bound on toast duration (seconds) for externally-triggered
- * toasts. Toasts are corner notifications so the cap is more generous than
- * Center Flash, but `0` (never auto-dismiss) is rejected - external scripts
- * that want a sticky toast must opt in explicitly via `dismissible: true`.
- */
-const EXTERNAL_TOAST_MAX_DURATION_SECONDS = 60;
-import { AGENT_IDS } from '../../../../shared/agentIds';
-import {
-	getActivePluginManager,
-	isPluginsFeatureEnabled,
-} from '../../../plugins/plugin-manager-singleton';
-import { evaluatePluginDispatch } from '../../../../shared/plugins/plugin-dispatch-gate';
-import {
-	CADENZA_OPS,
-	CADENZA_VIEW_TYPES,
-	CADENZA_COLORS,
-	type CadenzaPayload,
-	type CadenzaOp,
-	type CadenzaViewType,
-	type CadenzaColor,
-} from '../../../../shared/cadenza-types';
-import {
-	CONCERTO_CREATION_PHASES,
-	CONCERTO_PROGRESS_MAX_STEPS,
-	CONCERTO_PROGRESS_NOTE_VALUES,
-	MOVEMENT_OPS,
-	MOVEMENT_VIEW_TYPES,
-	type ConcertoCreationPhase,
-	type ConcertoProgressNote,
-	type ConcertoProgressNoteValue,
-	type MovementOp,
-	type MovementPayload,
-	type MovementStateSnapshot,
-	type MovementViewType,
-} from '../../../../shared/movement-types';
-import type {
-	ConcertoDesignerAction,
-	ConcertoDesignerActionResult,
-	MovementDesignerInspection,
-} from '../../../../shared/concerto-html';
 import { LOG_CONTEXT } from './shared';
 import type {
 	WebClientMessage,
@@ -136,6 +40,36 @@ import type {
 	MessageHandlerCallbacks,
 	MessageHandlerContext,
 } from './types';
+import { handleSendCommand, handleSwitchMode, handleSelectSession } from './commands';
+import {
+	handleGetSessions,
+	handleCreateSession,
+	handleCreateWorktreeSession,
+	handleDeleteSession,
+	handleRenameSession,
+	handleUpdateSessionCwd,
+	handleUpdateSessionSsh,
+	handleUpdateSessionConfig,
+	handleListDesktopSessions,
+	handleGetSessionHistory,
+} from './sessions';
+import { handleRefreshFileTree, handleGetFileTree } from './fileTree';
+import { handleNotifyToast, handleNotifyCenterFlash } from './notifications';
+import { handleCadenza } from './cadenza';
+import {
+	handleMovement,
+	handleGetMovementState,
+	handleGetMovementDesignerInspection,
+	handleInteractMovementDesigner,
+} from './movement';
+import { handleProfilingStart, handleProfilingStatus, handleProfilingStop } from './profiling';
+import {
+	handleMarketplaceGetManifest,
+	handleMarketplaceGetDocument,
+	handleMarketplaceGetReadme,
+	handleMarketplaceImportPlaybook,
+} from './marketplace';
+import { handlePluginsListTools, handlePluginsCallTool } from './plugins';
 import {
 	handleRefreshAutoRunDocs,
 	handleConfigureAutoRun,
@@ -163,6 +97,58 @@ import {
 	handleOpenTerminalTab,
 	handleNewAITabWithPrompt,
 } from './tabs';
+import { handleGetSettings, handleSetSetting } from './settings';
+import { handleTerminalWrite, handleTerminalResize } from './terminal';
+import {
+	handleGetGitStatus,
+	handleGetGitDiff,
+	handleGetGitBranches,
+	handleListWorktrees,
+} from './git';
+import {
+	handleGetGroups,
+	handleCreateGroup,
+	handleRenameGroup,
+	handleDeleteGroup,
+	handleMoveSessionToGroup,
+} from './groups';
+import {
+	handleGetGroupChats,
+	handleStartGroupChat,
+	handleGetGroupChatState,
+	handleSendGroupChatMessage,
+	handleStopGroupChat,
+} from './groupChat';
+import {
+	handleMergeContext,
+	handleTransferContext,
+	handleSummarizeContext,
+	handleCreateGist,
+} from './contextOps';
+import {
+	handleGetCueSubscriptions,
+	handleToggleCueSubscription,
+	handleGetCueActivity,
+	handleTriggerCueSubscription,
+	handleCuePipelineList,
+	handleCuePipelineGet,
+	handleCuePipelineSet,
+	handleCuePipelineRemove,
+} from './cue';
+import { handleEnqueueCommand, handleListQueue, handleRemoveQueueItem } from './queue';
+import {
+	handleListPlaybooks,
+	handleCreatePlaybook,
+	handleUpdatePlaybook,
+	handleDeletePlaybook,
+} from './playbooks';
+import {
+	handleGetUsageDashboard,
+	handleGetStatsAggregation,
+	handleStatsQuery,
+	handleGetAchievements,
+	handleGenerateDirectorNotesSynopsis,
+} from './stats';
 
 /**
  * WebSocket Message Handler Class
@@ -248,19 +234,19 @@ export class WebSocketMessageHandler {
 				break;
 
 			case 'send_command':
-				this.handleSendCommand(client, message);
+				handleSendCommand(this.ctx, client, message);
 				break;
 
 			case 'switch_mode':
-				this.handleSwitchMode(client, message);
+				handleSwitchMode(this.ctx, client, message);
 				break;
 
 			case 'select_session':
-				this.handleSelectSession(client, message);
+				handleSelectSession(this.ctx, client, message);
 				break;
 
 			case 'get_sessions':
-				this.handleGetSessions(client);
+				handleGetSessions(this.ctx, client);
 				break;
 
 			case 'get_app_info':
@@ -312,23 +298,23 @@ export class WebSocketMessageHandler {
 				break;
 
 			case 'enqueue_command':
-				this.handleEnqueueCommand(client, message);
+				handleEnqueueCommand(this.ctx, client, message);
 				break;
 
 			case 'list_queue':
-				this.handleListQueue(client, message);
+				handleListQueue(this.ctx, client, message);
 				break;
 
 			case 'remove_queue_item':
-				this.handleRemoveQueueItem(client, message);
+				handleRemoveQueueItem(this.ctx, client, message);
 				break;
 
 			case 'refresh_file_tree':
-				this.handleRefreshFileTree(client, message);
+				handleRefreshFileTree(this.ctx, client, message);
 				break;
 
 			case 'get_file_tree':
-				this.handleGetFileTree(client, message);
+				handleGetFileTree(this.ctx, client, message);
 				break;
 
 			case 'refresh_auto_run_docs':
@@ -340,7 +326,7 @@ export class WebSocketMessageHandler {
 				break;
 
 			case 'create_worktree_session':
-				this.handleCreateWorktreeSession(client, message);
+				handleCreateWorktreeSession(this.ctx, client, message);
 				break;
 
 			case 'set_auto_run_folder':
@@ -384,251 +370,251 @@ export class WebSocketMessageHandler {
 				break;
 
 			case 'list_playbooks':
-				this.handleListPlaybooks(client, message);
+				handleListPlaybooks(this.ctx, client, message);
 				break;
 
 			case 'create_playbook':
-				this.handleCreatePlaybook(client, message);
+				handleCreatePlaybook(this.ctx, client, message);
 				break;
 
 			case 'update_playbook':
-				this.handleUpdatePlaybook(client, message);
+				handleUpdatePlaybook(this.ctx, client, message);
 				break;
 
 			case 'delete_playbook':
-				this.handleDeletePlaybook(client, message);
+				handleDeletePlaybook(this.ctx, client, message);
 				break;
 
 			case 'get_settings':
-				this.handleGetSettings(client, message);
+				handleGetSettings(this.ctx, client, message);
 				break;
 
 			case 'set_setting':
-				this.handleSetSetting(client, message);
+				handleSetSetting(this.ctx, client, message);
 				break;
 
 			case 'create_session':
-				this.handleCreateSession(client, message);
+				handleCreateSession(this.ctx, client, message);
 				break;
 
 			case 'delete_session':
-				this.handleDeleteSession(client, message);
+				handleDeleteSession(this.ctx, client, message);
 				break;
 
 			case 'rename_session':
-				this.handleRenameSession(client, message);
+				handleRenameSession(this.ctx, client, message);
 				break;
 
 			case 'update_session_cwd':
-				this.handleUpdateSessionCwd(client, message);
+				handleUpdateSessionCwd(this.ctx, client, message);
 				break;
 
 			case 'update_session_ssh':
-				this.handleUpdateSessionSsh(client, message);
+				handleUpdateSessionSsh(this.ctx, client, message);
 				break;
 
 			case 'update_session_config':
-				this.handleUpdateSessionConfig(client, message);
+				handleUpdateSessionConfig(this.ctx, client, message);
 				break;
 
 			case 'get_groups':
-				this.handleGetGroups(client, message);
+				handleGetGroups(this.ctx, client, message);
 				break;
 
 			case 'create_group':
-				this.handleCreateGroup(client, message);
+				handleCreateGroup(this.ctx, client, message);
 				break;
 
 			case 'rename_group':
-				this.handleRenameGroup(client, message);
+				handleRenameGroup(this.ctx, client, message);
 				break;
 
 			case 'delete_group':
-				this.handleDeleteGroup(client, message);
+				handleDeleteGroup(this.ctx, client, message);
 				break;
 
 			case 'move_session_to_group':
-				this.handleMoveSessionToGroup(client, message);
+				handleMoveSessionToGroup(this.ctx, client, message);
 				break;
 
 			case 'get_git_status':
-				this.handleGetGitStatus(client, message);
+				handleGetGitStatus(this.ctx, client, message);
 				break;
 
 			case 'get_git_diff':
-				this.handleGetGitDiff(client, message);
+				handleGetGitDiff(this.ctx, client, message);
 				break;
 
 			case 'get_git_branches':
-				this.handleGetGitBranches(client, message);
+				handleGetGitBranches(this.ctx, client, message);
 				break;
 
 			case 'list_worktrees':
-				this.handleListWorktrees(client, message);
+				handleListWorktrees(this.ctx, client, message);
 				break;
 
 			case 'get_group_chats':
-				this.handleGetGroupChats(client, message);
+				handleGetGroupChats(this.ctx, client, message);
 				break;
 
 			case 'start_group_chat':
-				this.handleStartGroupChat(client, message);
+				handleStartGroupChat(this.ctx, client, message);
 				break;
 
 			case 'get_group_chat_state':
-				this.handleGetGroupChatState(client, message);
+				handleGetGroupChatState(this.ctx, client, message);
 				break;
 
 			case 'send_group_chat_message':
-				this.handleSendGroupChatMessage(client, message);
+				handleSendGroupChatMessage(this.ctx, client, message);
 				break;
 
 			case 'stop_group_chat':
-				this.handleStopGroupChat(client, message);
+				handleStopGroupChat(this.ctx, client, message);
 				break;
 
 			case 'merge_context':
-				this.handleMergeContext(client, message);
+				handleMergeContext(this.ctx, client, message);
 				break;
 
 			case 'transfer_context':
-				this.handleTransferContext(client, message);
+				handleTransferContext(this.ctx, client, message);
 				break;
 
 			case 'summarize_context':
-				this.handleSummarizeContext(client, message);
+				handleSummarizeContext(this.ctx, client, message);
 				break;
 
 			case 'create_gist':
-				this.handleCreateGist(client, message);
+				handleCreateGist(this.ctx, client, message);
 				break;
 
 			case 'get_cue_subscriptions':
-				this.handleGetCueSubscriptions(client, message);
+				handleGetCueSubscriptions(this.ctx, client, message);
 				break;
 
 			case 'toggle_cue_subscription':
-				this.handleToggleCueSubscription(client, message);
+				handleToggleCueSubscription(this.ctx, client, message);
 				break;
 
 			case 'get_cue_activity':
-				this.handleGetCueActivity(client, message);
+				handleGetCueActivity(this.ctx, client, message);
 				break;
 
 			case 'trigger_cue_subscription':
-				this.handleTriggerCueSubscription(client, message);
+				handleTriggerCueSubscription(this.ctx, client, message);
 				break;
 
 			case 'cue_pipeline_list':
-				this.handleCuePipelineList(client, message);
+				handleCuePipelineList(this.ctx, client, message);
 				break;
 
 			case 'cue_pipeline_get':
-				this.handleCuePipelineGet(client, message);
+				handleCuePipelineGet(this.ctx, client, message);
 				break;
 
 			case 'cue_pipeline_set':
-				this.handleCuePipelineSet(client, message);
+				handleCuePipelineSet(this.ctx, client, message);
 				break;
 
 			case 'cue_pipeline_remove':
-				this.handleCuePipelineRemove(client, message);
+				handleCuePipelineRemove(this.ctx, client, message);
 				break;
 
 			case 'get_usage_dashboard':
-				this.handleGetUsageDashboard(client, message);
+				handleGetUsageDashboard(this.ctx, client, message);
 				break;
 
 			case 'get_achievements':
-				this.handleGetAchievements(client, message);
+				handleGetAchievements(this.ctx, client, message);
 				break;
 
 			case 'get_stats_aggregation':
-				this.handleGetStatsAggregation(client, message);
+				handleGetStatsAggregation(this.ctx, client, message);
 				break;
 
 			case 'stats_query':
-				this.handleStatsQuery(client, message);
+				handleStatsQuery(this.ctx, client, message);
 				break;
 
 			case 'generate_director_notes_synopsis':
-				this.handleGenerateDirectorNotesSynopsis(client, message);
+				handleGenerateDirectorNotesSynopsis(this.ctx, client, message);
 				break;
 
 			case 'terminal_write':
-				this.handleTerminalWrite(client, message);
+				handleTerminalWrite(this.ctx, client, message);
 				break;
 
 			case 'terminal_resize':
-				this.handleTerminalResize(client, message);
+				handleTerminalResize(this.ctx, client, message);
 				break;
 
 			case 'notify_toast':
-				this.handleNotifyToast(client, message);
+				handleNotifyToast(this.ctx, client, message);
 				break;
 
 			case 'movement':
-				this.handleMovement(client, message);
+				handleMovement(this.ctx, client, message);
 				break;
 			case 'get_movement_state':
-				this.handleGetMovementState(client, message);
+				handleGetMovementState(this.ctx, client, message);
 				break;
 			case 'get_movement_designer_inspection':
-				this.handleGetMovementDesignerInspection(client, message);
+				handleGetMovementDesignerInspection(this.ctx, client, message);
 				break;
 			case 'interact_movement_designer':
-				this.handleInteractMovementDesigner(client, message);
+				handleInteractMovementDesigner(this.ctx, client, message);
 				break;
 			case 'cadenza':
-				this.handleCadenza(client, message);
+				handleCadenza(this.ctx, client, message);
 				break;
 
 			case 'notify_center_flash':
-				this.handleNotifyCenterFlash(client, message);
+				handleNotifyCenterFlash(this.ctx, client, message);
 				break;
 
 			case 'profiling_start':
-				void this.handleProfilingStart(client, message);
+				void handleProfilingStart(this.ctx, client, message);
 				break;
 
 			case 'profiling_stop':
-				void this.handleProfilingStop(client, message);
+				void handleProfilingStop(this.ctx, client, message);
 				break;
 
 			case 'profiling_status':
-				this.handleProfilingStatus(client, message);
+				handleProfilingStatus(this.ctx, client, message);
 				break;
 
 			case 'marketplace_get_manifest':
-				this.handleMarketplaceGetManifest(client, message);
+				handleMarketplaceGetManifest(this.ctx, client, message);
 				break;
 
 			case 'marketplace_get_document':
-				this.handleMarketplaceGetDocument(client, message);
+				handleMarketplaceGetDocument(this.ctx, client, message);
 				break;
 
 			case 'marketplace_get_readme':
-				this.handleMarketplaceGetReadme(client, message);
+				handleMarketplaceGetReadme(this.ctx, client, message);
 				break;
 
 			case 'marketplace_import_playbook':
-				this.handleMarketplaceImportPlaybook(client, message);
+				handleMarketplaceImportPlaybook(this.ctx, client, message);
 				break;
 
 			case 'list_desktop_sessions':
-				this.handleListDesktopSessions(client, message);
+				handleListDesktopSessions(this.ctx, client, message);
 				break;
 
 			case 'plugins_list_tools':
-				this.handlePluginsListTools(client, message);
+				handlePluginsListTools(this.ctx, client, message);
 				break;
 
 			case 'plugins_call_tool':
-				void this.handlePluginsCallTool(client, message);
+				void handlePluginsCallTool(this.ctx, client, message);
 				break;
 
 			case 'get_session_history':
-				this.handleGetSessionHistory(client, message);
+				handleGetSessionHistory(this.ctx, client, message);
 				break;
 
 			case 'bridge.invoke':
@@ -676,288 +662,6 @@ export class WebSocketMessageHandler {
 	}
 
 	/**
-	 * Handle send_command message - execute command in session
-	 */
-	private handleSendCommand(client: WebClient, message: WebClientMessage): void {
-		const sessionId = message.sessionId as string;
-		const command = message.command as string;
-		// inputMode from web client - use this instead of server state to avoid sync issues
-		const clientInputMode = message.inputMode as 'ai' | 'terminal' | undefined;
-		// Optional explicit tab target. When omitted, the renderer falls back to
-		// the active tab (legacy `send --live` behavior). Used by
-		// `maestro-cli dispatch --session <tabId>` to address a specific tab.
-		const requestedTabId = typeof message.tabId === 'string' ? message.tabId : undefined;
-		// force=true bypasses the busy-state guard below, allowing callers to
-		// dispatch concurrent writes to an already-running agent. Used by
-		// `maestro-cli dispatch --force`.
-		const force = message.force === true;
-		// background=true suppresses the desktop's focus side effect (switching
-		// to the target agent/tab). `maestro-cli dispatch` sets this by default;
-		// passing `--focus` clears it to bring the target to the foreground.
-		const background = message.background === true;
-		// Optional base64 data URLs pasted from the web client. Threaded through
-		// to the renderer so AI tabs can include them in the agent prompt.
-		const images = Array.isArray(message.images)
-			? (message.images as unknown[]).filter((v): v is string => typeof v === 'string')
-			: undefined;
-
-		logger.info(
-			`[Web Command] Received: sessionId=${sessionId}, inputMode=${clientInputMode}, command=${command?.substring(0, 50)}, images=${images?.length ?? 0}`,
-			LOG_CONTEXT
-		);
-
-		// Image-only sends are valid in AI mode (the composer lets users paste
-		// images and submit without typing), so the guard accepts either a
-		// non-empty command OR at least one image. Normalize a missing command
-		// to '' so the renderer's downstream LogEntry.text stays a string.
-		const hasImages = !!images && images.length > 0;
-		if (!sessionId || (!command && !hasImages)) {
-			logger.warn(
-				`[Web Command] Missing sessionId or command/images: sessionId=${sessionId}, commandLen=${command?.length}, images=${images?.length ?? 0}`,
-				LOG_CONTEXT
-			);
-			this.sendError(client, 'Missing sessionId or command');
-			return;
-		}
-		const effectiveCommand = command ?? '';
-
-		// Get session details to check state and determine how to handle
-		const sessionDetail = this.callbacks.getSessionDetail?.(sessionId);
-		if (!sessionDetail) {
-			this.sendError(client, 'Session not found');
-			return;
-		}
-
-		// Check if session is busy - prevent race conditions between desktop and web.
-		// `force: true` opts out of this guard (see `maestro-cli send --live --force`).
-		if (sessionDetail.state === 'busy' && !force) {
-			this.sendError(
-				client,
-				'Session is busy - please wait for the current operation to complete',
-				{
-					sessionId,
-				}
-			);
-			logger.debug(`Command rejected - session ${sessionId} is busy`, LOG_CONTEXT);
-			return;
-		}
-		if (sessionDetail.state === 'busy' && force) {
-			logger.info(`[Web Command] Force-dispatching to busy session ${sessionId}`, LOG_CONTEXT);
-		}
-
-		// Use client's inputMode if provided, otherwise fall back to server state
-		const effectiveMode = clientInputMode || sessionDetail.inputMode;
-		const isAiMode = effectiveMode === 'ai';
-		const mode = isAiMode ? 'AI' : 'CLI';
-		const claudeId = sessionDetail.agentSessionId || 'none';
-
-		// Log all web interface commands prominently
-		logger.info(
-			`[Web Command] Mode: ${mode} | Session: ${sessionId}${isAiMode ? ` | Claude: ${claudeId}` : ''} | Message: ${effectiveCommand} | Images: ${images?.length ?? 0}`,
-			LOG_CONTEXT
-		);
-
-		// Only echo a tabId in command_result when the caller passed one
-		// explicitly. Returning the server's snapshot of `activeTabId` for the
-		// no-tabId path would lie when the user switches active tabs between
-		// the IPC send and IPC receive - callers chaining `dispatch --session
-		// <returnedTabId>` would think they are continuing a conversation that
-		// actually went to a different tab. For deterministic addressing,
-		// callers should use `dispatch --new-tab` (returns the new tabId from
-		// the renderer ack) and then `dispatch --session <tabId>` (echoes back
-		// the caller-supplied authoritative tabId).
-		const resolvedTabId = requestedTabId;
-
-		// Route ALL commands through the renderer for consistent handling
-		// The renderer handles both AI and terminal modes, updating UI and state
-		// Pass clientInputMode so renderer uses the web's intended mode
-		if (this.callbacks.executeCommand) {
-			this.callbacks
-				.executeCommand(
-					sessionId,
-					effectiveCommand,
-					clientInputMode,
-					requestedTabId,
-					force,
-					images,
-					background
-				)
-				.then((success) => {
-					this.send(client, {
-						type: 'command_result',
-						success,
-						sessionId,
-						...(resolvedTabId ? { tabId: resolvedTabId } : {}),
-						requestId: message.requestId,
-					});
-					if (!success) {
-						logger.warn(
-							`[Web Command] ${mode} command rejected for session ${sessionId}`,
-							LOG_CONTEXT
-						);
-					}
-				})
-				.catch((error) => {
-					logger.error(
-						`[Web Command] ${mode} command failed for session ${sessionId}: ${error.message}`,
-						LOG_CONTEXT
-					);
-					this.sendError(client, `Failed to execute command: ${error.message}`);
-				});
-		} else {
-			this.sendError(client, 'Command execution not configured');
-		}
-	}
-
-	/**
-	 * Handle switch_mode message - switch between AI and terminal mode
-	 *
-	 * When switching to terminal mode, spawns a dedicated PTY process for the web client
-	 * (session ID: {sessionId}-terminal). When switching back to AI, kills it.
-	 */
-	private handleSwitchMode(client: WebClient, message: WebClientMessage): void {
-		const sessionId = message.sessionId as string;
-		const mode = message.mode as 'ai' | 'terminal';
-		logger.info(
-			`[Web] Received switch_mode message: session=${sessionId}, mode=${mode}`,
-			LOG_CONTEXT
-		);
-
-		if (!sessionId || !mode) {
-			this.sendError(client, 'Missing sessionId or mode');
-			return;
-		}
-
-		if (!this.callbacks.switchMode) {
-			logger.warn(`[Web] switchModeCallback is not set!`, LOG_CONTEXT);
-			this.sendError(client, 'Mode switching not configured');
-			return;
-		}
-
-		// Forward to desktop's mode switching logic
-		// This ensures single source of truth - desktop handles state updates and broadcasts
-		logger.info(`[Web] Calling switchModeCallback for session ${sessionId}: ${mode}`, LOG_CONTEXT);
-		this.callbacks
-			.switchMode(sessionId, mode)
-			.then(async (success) => {
-				// Spawn or kill the web terminal PTY based on mode
-				if (success && mode === 'terminal') {
-					// Look up session CWD for the terminal working directory
-					const sessionDetail = this.callbacks.getSessionDetail?.(sessionId);
-					const cwd = sessionDetail?.cwd || process.cwd();
-					try {
-						const spawnResult = await this.callbacks.spawnTerminalForWeb?.(sessionId, { cwd });
-						logger.info(
-							`[Web] Terminal PTY spawn for ${sessionId}: success=${spawnResult?.success}`,
-							LOG_CONTEXT
-						);
-						if (spawnResult?.success) {
-							// Notify the web client that the PTY is ready so it can re-send
-							// its current dimensions (the initial resize fired before the PTY existed)
-							this.send(client, {
-								type: 'terminal_ready',
-								sessionId,
-							});
-						} else {
-							// PTY failed to spawn - report failure so the client can roll back
-							this.send(client, {
-								type: 'mode_switch_result',
-								success: false,
-								sessionId,
-								mode,
-								error: 'Failed to spawn terminal PTY',
-								requestId: message.requestId,
-							});
-							return;
-						}
-					} catch (err) {
-						logger.error(
-							`[Web] Failed to spawn terminal PTY for ${sessionId}: ${err}`,
-							LOG_CONTEXT
-						);
-						this.send(client, {
-							type: 'mode_switch_result',
-							success: false,
-							sessionId,
-							mode,
-							error: `Failed to spawn terminal: ${err instanceof Error ? err.message : String(err)}`,
-							requestId: message.requestId,
-						});
-						return;
-					}
-				}
-				// When switching back to AI, keep the terminal PTY alive so the user
-				// can return to a running process (e.g. npm run dev). The PTY is only
-				// killed when the session itself is removed.
-
-				this.send(client, {
-					type: 'mode_switch_result',
-					success,
-					sessionId,
-					mode,
-					requestId: message.requestId,
-				});
-				logger.debug(
-					`Mode switch for session ${sessionId} to ${mode}: ${success ? 'success' : 'failed'}`,
-					LOG_CONTEXT
-				);
-			})
-			.catch((error) => {
-				this.sendError(client, `Failed to switch mode: ${error.message}`);
-			});
-	}
-
-	/**
-	 * Handle select_session message - select/switch to a session in desktop
-	 */
-	private handleSelectSession(client: WebClient, message: WebClientMessage): void {
-		const sessionId = message.sessionId as string;
-		const tabId = message.tabId as string | undefined;
-		const focus = message.focus as boolean | undefined;
-		logger.info(
-			`[Web] Received select_session message: session=${sessionId}, tab=${tabId || 'none'}, focus=${focus || false}`,
-			LOG_CONTEXT
-		);
-
-		if (!sessionId) {
-			this.sendError(client, 'Missing sessionId');
-			return;
-		}
-
-		if (!this.callbacks.selectSession) {
-			logger.warn(`[Web] selectSessionCallback is not set!`, LOG_CONTEXT);
-			this.sendError(client, 'Session selection not configured');
-			return;
-		}
-
-		// Forward to desktop's session selection logic (include tabId if provided)
-		logger.info(
-			`[Web] Calling selectSessionCallback for session ${sessionId}${tabId ? `, tab ${tabId}` : ''}`,
-			LOG_CONTEXT
-		);
-		this.callbacks
-			.selectSession(sessionId, tabId, focus)
-			.then((success) => {
-				if (success) {
-					// Subscribe client to this session's output so they receive session_output messages
-					client.subscribedSessionId = sessionId;
-					logger.debug(`Session ${sessionId} selected in desktop, client subscribed`, LOG_CONTEXT);
-				} else {
-					logger.warn(`Failed to select session ${sessionId} in desktop`, LOG_CONTEXT);
-				}
-				this.send(client, {
-					type: 'select_session_result',
-					success,
-					sessionId,
-					requestId: message.requestId,
-				});
-			})
-			.catch((error) => {
-				this.sendError(client, `Failed to select session: ${error.message}`);
-			});
-	}
-
-	/**
 	 * Handle get_sessions message - request updated sessions list
 	 */
 	/**
@@ -972,3476 +676,6 @@ export class WebSocketMessageHandler {
 			version: app.getVersion(),
 			commitHash: getCommitHash(),
 			platform: process.platform,
-		});
-	}
-
-	private handleGetSessions(client: WebClient): void {
-		if (
-			this.callbacks.getSessions &&
-			this.callbacks.getLiveSessionInfo &&
-			this.callbacks.isSessionLive
-		) {
-			const allSessions = this.callbacks.getSessions();
-			// Enrich sessions with live info if available
-			const sessionsWithLiveInfo = allSessions.map((s) => {
-				const liveInfo = this.callbacks.getLiveSessionInfo!(s.id);
-				return {
-					...s,
-					agentSessionId: liveInfo?.agentSessionId || s.agentSessionId,
-					liveEnabledAt: liveInfo?.enabledAt,
-					isLive: this.callbacks.isSessionLive!(s.id),
-				};
-			});
-			this.send(client, { type: 'sessions_list', sessions: sessionsWithLiveInfo });
-		}
-	}
-
-	/**
-	 * Handle refresh_file_tree message - refresh the file tree for a session
-	 */
-	private handleRefreshFileTree(client: WebClient, message: WebClientMessage): void {
-		const sessionId = message.sessionId as string;
-		logger.info(`[Web] Received refresh_file_tree message: session=${sessionId}`, LOG_CONTEXT);
-
-		if (!sessionId) {
-			this.sendError(client, 'Missing sessionId');
-			return;
-		}
-
-		if (!this.callbacks.refreshFileTree) {
-			this.sendError(client, 'File tree refresh not configured');
-			return;
-		}
-
-		this.callbacks
-			.refreshFileTree(sessionId)
-			.then((success) => {
-				this.send(client, {
-					type: 'refresh_file_tree_result',
-					success,
-					sessionId,
-					requestId: message.requestId,
-				});
-			})
-			.catch((error) => {
-				this.sendError(client, `Failed to refresh file tree: ${error.message}`);
-			});
-	}
-
-	/**
-	 * Handle get_file_tree message - read directory tree for file explorer
-	 * Uses Node.js fs directly (no IPC to renderer needed)
-	 */
-	private handleGetFileTree(client: WebClient, message: WebClientMessage): void {
-		const sessionId = message.sessionId as string;
-		const dirPath = message.path as string;
-		const maxDepth = Math.min((message.maxDepth as number) || 3, 5);
-
-		if (!dirPath) {
-			this.sendError(client, 'Missing path for get_file_tree');
-			return;
-		}
-
-		// Validate dirPath is within the session's working directory
-		const sessionDetail = this.callbacks.getSessionDetail?.(sessionId);
-		if (!sessionDetail?.cwd) {
-			this.sendError(client, 'Cannot resolve session working directory');
-			return;
-		}
-		const resolvedDir = path.resolve(dirPath);
-		const resolvedCwd = path.resolve(sessionDetail.cwd);
-		if (!resolvedDir.startsWith(resolvedCwd + path.sep) && resolvedDir !== resolvedCwd) {
-			this.sendError(client, 'Requested path is outside the session working directory');
-			return;
-		}
-
-		this.buildFileTree(dirPath, maxDepth)
-			.then((tree) => {
-				this.send(client, {
-					type: 'file_tree_data',
-					sessionId,
-					tree,
-					path: dirPath,
-					requestId: message.requestId,
-				});
-			})
-			.catch((error) => {
-				this.send(client, {
-					type: 'file_tree_data',
-					sessionId,
-					tree: [],
-					error: error.message,
-					path: dirPath,
-					requestId: message.requestId,
-				});
-			});
-	}
-
-	/**
-	 * Recursively build a file tree from a directory path
-	 */
-	private async buildFileTree(
-		dirPath: string,
-		maxDepth: number,
-		currentDepth = 0
-	): Promise<
-		Array<{
-			name: string;
-			type: 'file' | 'folder';
-			children?: Array<{ name: string; type: 'file' | 'folder'; children?: any[]; path: string }>;
-			path: string;
-		}>
-	> {
-		// Common ignore patterns
-		const IGNORE = new Set([
-			'node_modules',
-			'.git',
-			'.next',
-			'.nuxt',
-			'dist',
-			'build',
-			'.cache',
-			'__pycache__',
-			'.tox',
-			'.mypy_cache',
-			'.pytest_cache',
-			'venv',
-			'.venv',
-			'target',
-			'.idea',
-			'.vscode',
-			'.DS_Store',
-			'Thumbs.db',
-			'.turbo',
-			'coverage',
-			'.nyc_output',
-			'.parcel-cache',
-			'.svelte-kit',
-		]);
-
-		try {
-			const entries = await fs.readdir(dirPath, { withFileTypes: true });
-			const result: Array<{
-				name: string;
-				type: 'file' | 'folder';
-				children?: any[];
-				path: string;
-			}> = [];
-
-			// Sort: folders first, then alphabetically
-			const sorted = entries
-				.filter((e) => !IGNORE.has(e.name) && !e.name.startsWith('.'))
-				.sort((a, b) => {
-					if (a.isDirectory() && !b.isDirectory()) return -1;
-					if (!a.isDirectory() && b.isDirectory()) return 1;
-					return a.name.localeCompare(b.name);
-				});
-
-			for (const entry of sorted) {
-				const fullPath = path.join(dirPath, entry.name);
-
-				if (entry.isDirectory()) {
-					const children =
-						currentDepth < maxDepth
-							? await this.buildFileTree(fullPath, maxDepth, currentDepth + 1)
-							: undefined;
-					result.push({
-						name: entry.name,
-						type: 'folder',
-						children,
-						path: fullPath,
-					});
-				} else {
-					result.push({
-						name: entry.name,
-						type: 'file',
-						path: fullPath,
-					});
-				}
-			}
-
-			return result;
-		} catch {
-			// Permission denied or other errors - return empty
-			return [];
-		}
-	}
-
-	/**
-	 * Handle enqueue_command message - hand a CLI prompt to the renderer's
-	 * authoritative execution queue. When the target session is busy the prompt
-	 * is appended to `session.executionQueue` (FIFO); when idle it is dispatched
-	 * immediately through the same path as a plain send. Used by
-	 * `maestro-cli dispatch --queue`. Unlike `send_command`, this never rejects a
-	 * busy target: waiting in line is the whole point of the mode.
-	 */
-	private handleEnqueueCommand(client: WebClient, message: WebClientMessage): void {
-		const sessionId = typeof message.sessionId === 'string' ? message.sessionId : '';
-		const command = typeof message.command === 'string' ? message.command : '';
-		const clientInputMode = message.inputMode as 'ai' | 'terminal' | undefined;
-		const requestedTabId = typeof message.tabId === 'string' ? message.tabId : undefined;
-		const background = message.background === true;
-		const images = Array.isArray(message.images)
-			? (message.images as unknown[]).filter((v): v is string => typeof v === 'string')
-			: undefined;
-
-		logger.info(
-			`[Web] Received enqueue_command: session=${sessionId}, tab=${requestedTabId ?? 'active'}, commandLen=${command.length}, images=${images?.length ?? 0}`,
-			LOG_CONTEXT
-		);
-
-		const sendErrorResult = (error: string) => {
-			this.send(client, {
-				type: 'enqueue_command_result',
-				success: false,
-				error,
-				sessionId,
-				requestId: message.requestId,
-			});
-		};
-
-		const hasImages = !!images && images.length > 0;
-		if (!sessionId || (!command && !hasImages)) {
-			sendErrorResult('Missing sessionId or command');
-			return;
-		}
-
-		if (!this.callbacks.getSessionDetail?.(sessionId)) {
-			sendErrorResult('Session not found');
-			return;
-		}
-
-		if (!this.callbacks.enqueueCommand) {
-			sendErrorResult('Enqueue not configured');
-			return;
-		}
-
-		this.callbacks
-			.enqueueCommand(sessionId, command ?? '', clientInputMode, requestedTabId, images, background)
-			.then((result) => {
-				this.send(client, {
-					type: 'enqueue_command_result',
-					success: result.success,
-					sessionId,
-					...(result.tabId ? { tabId: result.tabId } : {}),
-					...(result.queued !== undefined ? { queued: result.queued } : {}),
-					...(result.queuePosition !== undefined ? { queuePosition: result.queuePosition } : {}),
-					...(result.queueLength !== undefined ? { queueLength: result.queueLength } : {}),
-					...(result.itemId ? { itemId: result.itemId } : {}),
-					...(result.error ? { error: result.error } : {}),
-					requestId: message.requestId,
-				});
-			})
-			.catch((error) => {
-				captureException(error instanceof Error ? error : new Error(String(error)), {
-					extra: { area: 'web-server', handler: 'enqueue_command', sessionId },
-				});
-				sendErrorResult(`Failed to enqueue command: ${error.message}`);
-			});
-	}
-
-	/**
-	 * Handle list_queue message - return a snapshot of the renderer's execution
-	 * queue(s). Used by `maestro-cli queue list`. When sessionId is omitted, every
-	 * session with queued items is returned.
-	 */
-	private handleListQueue(client: WebClient, message: WebClientMessage): void {
-		const sessionId = typeof message.sessionId === 'string' ? message.sessionId : undefined;
-		if (!this.callbacks.listQueue) {
-			this.send(client, {
-				type: 'list_queue_result',
-				success: false,
-				queues: [],
-				error: 'List queue not configured',
-				requestId: message.requestId,
-			});
-			return;
-		}
-		this.callbacks
-			.listQueue(sessionId)
-			.then((result) => {
-				this.send(client, {
-					type: 'list_queue_result',
-					success: result.success,
-					queues: result.queues,
-					...(result.error ? { error: result.error } : {}),
-					requestId: message.requestId,
-				});
-			})
-			.catch((error) => {
-				captureException(error instanceof Error ? error : new Error(String(error)), {
-					extra: { area: 'web-server', handler: 'list_queue', sessionId },
-				});
-				this.send(client, {
-					type: 'list_queue_result',
-					success: false,
-					queues: [],
-					error: `Failed to list queue: ${error.message}`,
-					requestId: message.requestId,
-				});
-			});
-	}
-
-	/**
-	 * Handle remove_queue_item message - drop a queued item by id from the
-	 * renderer's execution queue. Used by `maestro-cli queue remove`.
-	 */
-	private handleRemoveQueueItem(client: WebClient, message: WebClientMessage): void {
-		const sessionId = typeof message.sessionId === 'string' ? message.sessionId : '';
-		const itemId = typeof message.itemId === 'string' ? message.itemId : '';
-		if (!sessionId || !itemId) {
-			this.send(client, {
-				type: 'remove_queue_item_result',
-				success: false,
-				removed: false,
-				error: 'Missing sessionId or itemId',
-				requestId: message.requestId,
-			});
-			return;
-		}
-		if (!this.callbacks.removeQueueItem) {
-			this.send(client, {
-				type: 'remove_queue_item_result',
-				success: false,
-				removed: false,
-				error: 'Remove queue item not configured',
-				requestId: message.requestId,
-			});
-			return;
-		}
-		this.callbacks
-			.removeQueueItem(sessionId, itemId)
-			.then((result) => {
-				this.send(client, {
-					type: 'remove_queue_item_result',
-					success: result.success,
-					removed: result.removed,
-					...(result.error ? { error: result.error } : {}),
-					requestId: message.requestId,
-				});
-			})
-			.catch((error) => {
-				captureException(error instanceof Error ? error : new Error(String(error)), {
-					extra: { area: 'web-server', handler: 'remove_queue_item', sessionId, itemId },
-				});
-				this.send(client, {
-					type: 'remove_queue_item_result',
-					success: false,
-					removed: false,
-					error: `Failed to remove queue item: ${error.message}`,
-					requestId: message.requestId,
-				});
-			});
-	}
-
-	/**
-	 * Handle list_playbooks message - return the saved playbooks for a session.
-	 */
-	private handleListPlaybooks(client: WebClient, message: WebClientMessage): void {
-		const sessionId = message.sessionId as string;
-		if (!sessionId) {
-			this.sendError(client, 'Missing sessionId');
-			return;
-		}
-		if (!this.callbacks.listPlaybooks) {
-			this.sendError(client, 'Playbook listing not configured');
-			return;
-		}
-		this.callbacks
-			.listPlaybooks(sessionId)
-			.then((playbooks) => {
-				this.send(client, {
-					type: 'playbooks_list',
-					sessionId,
-					playbooks,
-					requestId: message.requestId,
-				});
-			})
-			.catch((error) => {
-				this.reportHandlerError(
-					client,
-					error,
-					'list_playbooks',
-					{ sessionId, requestId: message.requestId },
-					'Failed to list playbooks'
-				);
-			});
-	}
-
-	/**
-	 * Handle create_playbook message - persist a new playbook with the given config.
-	 */
-	private handleCreatePlaybook(client: WebClient, message: WebClientMessage): void {
-		const sessionId = message.sessionId as string;
-		const playbook = message.playbook as
-			| {
-					name?: unknown;
-					documents?: unknown;
-					loopEnabled?: unknown;
-					maxLoops?: unknown;
-					prompt?: unknown;
-			  }
-			| undefined;
-
-		if (!sessionId) {
-			this.sendError(client, 'Missing sessionId');
-			return;
-		}
-		if (!playbook || typeof playbook !== 'object') {
-			this.sendError(client, 'Missing playbook payload');
-			return;
-		}
-		if (typeof playbook.name !== 'string' || playbook.name.trim() === '') {
-			this.sendError(client, 'Playbook name must be a non-empty string');
-			return;
-		}
-		const documents = this.parsePlaybookDocuments(playbook.documents);
-		if (documents === null) {
-			this.sendError(client, 'Invalid playbook documents');
-			return;
-		}
-		if (playbook.loopEnabled !== undefined && typeof playbook.loopEnabled !== 'boolean') {
-			this.sendError(client, 'loopEnabled must be a boolean');
-			return;
-		}
-		if (
-			playbook.maxLoops !== undefined &&
-			playbook.maxLoops !== null &&
-			(typeof playbook.maxLoops !== 'number' ||
-				!Number.isFinite(playbook.maxLoops) ||
-				playbook.maxLoops < 0)
-		) {
-			this.sendError(client, 'maxLoops must be a finite non-negative number');
-			return;
-		}
-		if (playbook.prompt !== undefined && typeof playbook.prompt !== 'string') {
-			this.sendError(client, 'prompt must be a string');
-			return;
-		}
-
-		if (!this.callbacks.createPlaybook) {
-			this.sendError(client, 'Playbook creation not configured');
-			return;
-		}
-
-		this.callbacks
-			.createPlaybook(sessionId, {
-				name: playbook.name.trim(),
-				documents,
-				loopEnabled: Boolean(playbook.loopEnabled),
-				maxLoops: (playbook.maxLoops as number | null | undefined) ?? null,
-				prompt: typeof playbook.prompt === 'string' ? playbook.prompt : '',
-			})
-			.then((created) => {
-				this.send(client, {
-					type: 'create_playbook_result',
-					success: created !== null,
-					sessionId,
-					playbook: created,
-					requestId: message.requestId,
-				});
-			})
-			.catch((error) => {
-				this.reportHandlerError(
-					client,
-					error,
-					'create_playbook',
-					{ sessionId, requestId: message.requestId },
-					'Failed to create playbook'
-				);
-			});
-	}
-
-	/**
-	 * Handle update_playbook message - apply partial updates to an existing playbook.
-	 */
-	private handleUpdatePlaybook(client: WebClient, message: WebClientMessage): void {
-		const sessionId = message.sessionId as string;
-		const playbookId = message.playbookId as string;
-		const updates = message.updates as
-			| {
-					name?: unknown;
-					documents?: unknown;
-					loopEnabled?: unknown;
-					maxLoops?: unknown;
-					prompt?: unknown;
-			  }
-			| undefined;
-
-		if (!sessionId || !playbookId) {
-			this.sendError(client, 'Missing sessionId or playbookId');
-			return;
-		}
-		if (!updates || typeof updates !== 'object') {
-			this.sendError(client, 'Missing updates payload');
-			return;
-		}
-
-		const sanitized: Partial<{
-			name: string;
-			documents: WebPlaybookDocument[];
-			loopEnabled: boolean;
-			maxLoops: number | null;
-			prompt: string;
-		}> = {};
-
-		if (updates.name !== undefined) {
-			if (typeof updates.name !== 'string' || updates.name.trim() === '') {
-				this.sendError(client, 'Playbook name must be a non-empty string');
-				return;
-			}
-			sanitized.name = updates.name.trim();
-		}
-		if (updates.documents !== undefined) {
-			const docs = this.parsePlaybookDocuments(updates.documents);
-			if (docs === null) {
-				this.sendError(client, 'Invalid playbook documents');
-				return;
-			}
-			sanitized.documents = docs;
-		}
-		if (updates.loopEnabled !== undefined) {
-			if (typeof updates.loopEnabled !== 'boolean') {
-				this.sendError(client, 'loopEnabled must be a boolean');
-				return;
-			}
-			sanitized.loopEnabled = updates.loopEnabled;
-		}
-		if (updates.maxLoops !== undefined) {
-			if (
-				updates.maxLoops !== null &&
-				(typeof updates.maxLoops !== 'number' ||
-					!Number.isFinite(updates.maxLoops) ||
-					updates.maxLoops < 0)
-			) {
-				this.sendError(client, 'maxLoops must be a finite non-negative number');
-				return;
-			}
-			sanitized.maxLoops = updates.maxLoops as number | null;
-		}
-		if (updates.prompt !== undefined) {
-			if (typeof updates.prompt !== 'string') {
-				this.sendError(client, 'prompt must be a string');
-				return;
-			}
-			sanitized.prompt = updates.prompt;
-		}
-
-		if (!this.callbacks.updatePlaybook) {
-			this.sendError(client, 'Playbook updates not configured');
-			return;
-		}
-
-		this.callbacks
-			.updatePlaybook(sessionId, playbookId, sanitized)
-			.then((updated) => {
-				this.send(client, {
-					type: 'update_playbook_result',
-					success: updated !== null,
-					sessionId,
-					playbook: updated,
-					requestId: message.requestId,
-				});
-			})
-			.catch((error) => {
-				this.reportHandlerError(
-					client,
-					error,
-					'update_playbook',
-					{ sessionId, playbookId, requestId: message.requestId },
-					'Failed to update playbook'
-				);
-			});
-	}
-
-	/**
-	 * Handle delete_playbook message - remove a playbook.
-	 */
-	private handleDeletePlaybook(client: WebClient, message: WebClientMessage): void {
-		const sessionId = message.sessionId as string;
-		const playbookId = message.playbookId as string;
-		if (!sessionId || !playbookId) {
-			this.sendError(client, 'Missing sessionId or playbookId');
-			return;
-		}
-		if (!this.callbacks.deletePlaybook) {
-			this.sendError(client, 'Playbook deletion not configured');
-			return;
-		}
-		this.callbacks
-			.deletePlaybook(sessionId, playbookId)
-			.then((success) => {
-				this.send(client, {
-					type: 'delete_playbook_result',
-					success,
-					sessionId,
-					playbookId,
-					requestId: message.requestId,
-				});
-			})
-			.catch((error) => {
-				this.reportHandlerError(
-					client,
-					error,
-					'delete_playbook',
-					{ sessionId, playbookId, requestId: message.requestId },
-					'Failed to delete playbook'
-				);
-			});
-	}
-
-	/**
-	 * Validate and normalize the `documents` field of a playbook payload.
-	 * Returns the parsed array on success, or null on any validation failure.
-	 *
-	 * Filenames may contain forward-slash subdirectories (e.g. `loop/step-1`)
-	 * but must not:
-	 *   - contain `..` path-traversal segments
-	 *   - contain backslashes (we only persist POSIX separators)
-	 *   - be absolute (POSIX `/foo` or Windows drive-letter `C:/foo`)
-	 *
-	 * `resetOnCompletion` is validated strictly as a boolean when present;
-	 * any other type is rejected rather than coerced, so a stray truthy
-	 * value from a buggy client can't silently flip the flag on.
-	 */
-	private parsePlaybookDocuments(input: unknown): WebPlaybookDocument[] | null {
-		if (!Array.isArray(input)) return null;
-		const out: WebPlaybookDocument[] = [];
-		for (const entry of input) {
-			if (!entry || typeof entry !== 'object') return null;
-			const e = entry as { filename?: unknown; resetOnCompletion?: unknown };
-			if (typeof e.filename !== 'string' || e.filename.trim() === '') return null;
-			if (e.filename.includes('..') || e.filename.includes('\\')) return null;
-			if (e.filename.startsWith('/')) return null;
-			if (/^[A-Za-z]:[\\/]/.test(e.filename)) return null;
-			let resetOnCompletion = false;
-			if (e.resetOnCompletion !== undefined) {
-				if (typeof e.resetOnCompletion !== 'boolean') return null;
-				resetOnCompletion = e.resetOnCompletion;
-			}
-			out.push({
-				filename: e.filename,
-				resetOnCompletion,
-			});
-		}
-		return out;
-	}
-
-	/**
-	 * Allowlist of setting keys modifiable from the web interface.
-	 */
-	private static readonly ALLOWED_SETTING_KEYS = new Set([
-		'activeThemeId',
-		'customThemeColors',
-		'customThemeBaseId',
-		'fontSize',
-		'enterToSendAI',
-		'defaultSaveToHistory',
-		'defaultShowThinking',
-		'notificationsEnabled',
-		'audioFeedbackEnabled',
-		'colorBlindMode',
-		'conductorProfile',
-		'maxOutputLines',
-		'encoreFeatures',
-	]);
-
-	/**
-	 * Handle get_settings message - return current settings
-	 */
-	private handleGetSettings(client: WebClient, message: WebClientMessage): void {
-		if (!this.callbacks.getSettings) {
-			this.sendError(client, 'Settings not configured');
-			return;
-		}
-
-		const settings = this.callbacks.getSettings();
-		this.send(client, {
-			type: 'settings',
-			settings,
-			requestId: message.requestId,
-		});
-	}
-
-	/**
-	 * Handle set_setting message - modify a single setting
-	 */
-	private handleSetSetting(client: WebClient, message: WebClientMessage): void {
-		const key = message.key as string;
-		const value = message.value as SettingValue;
-
-		if (!key || typeof key !== 'string') {
-			this.sendError(client, 'Missing or invalid setting key');
-			return;
-		}
-
-		if (!WebSocketMessageHandler.ALLOWED_SETTING_KEYS.has(key)) {
-			this.sendError(client, `Setting key '${key}' is not modifiable from the web interface`);
-			return;
-		}
-
-		if (value === undefined) {
-			this.sendError(client, 'Missing setting value');
-			return;
-		}
-
-		if (!this.callbacks.setSetting) {
-			this.sendError(client, 'Setting modification not configured');
-			return;
-		}
-
-		this.callbacks
-			.setSetting(key, value)
-			.then((success) => {
-				this.send(client, {
-					type: 'set_setting_result',
-					success,
-					key,
-					requestId: message.requestId,
-				});
-			})
-			.catch((error) => {
-				this.sendError(client, `Failed to set setting: ${error.message}`);
-			});
-	}
-
-	/**
-	 * Known agent types for validation - derived from the canonical AGENT_IDS list.
-	 * Excludes 'terminal' since it's an internal-only agent type.
-	 */
-	private static readonly VALID_AGENT_TYPES: Set<string> = new Set(
-		AGENT_IDS.filter((id) => id !== 'terminal')
-	);
-
-	/**
-	 * Handle create_session message - create a new agent session
-	 */
-	private handleCreateSession(client: WebClient, message: WebClientMessage): void {
-		const name = message.name as string;
-		const toolType = message.toolType as string;
-		const cwd = message.cwd as string;
-		const groupId = message.groupId as string | undefined;
-
-		if (!name || typeof name !== 'string') {
-			this.sendError(client, 'Missing or invalid name');
-			return;
-		}
-
-		if (!toolType || !WebSocketMessageHandler.VALID_AGENT_TYPES.has(toolType)) {
-			this.sendError(
-				client,
-				`Invalid toolType. Must be one of: ${[...WebSocketMessageHandler.VALID_AGENT_TYPES].join(', ')}`
-			);
-			return;
-		}
-
-		if (!cwd || typeof cwd !== 'string') {
-			this.sendError(client, 'Missing or invalid cwd');
-			return;
-		}
-
-		if (!this.callbacks.createSession) {
-			this.sendError(client, 'Session creation not configured');
-			return;
-		}
-
-		// Extract optional config fields
-		const config: CreateSessionConfig = {};
-		if (message.nudgeMessage) config.nudgeMessage = message.nudgeMessage as string;
-		if (message.newSessionMessage) config.newSessionMessage = message.newSessionMessage as string;
-		if (message.customPath) config.customPath = message.customPath as string;
-		if (message.customArgs) config.customArgs = message.customArgs as string;
-		if (message.customEnvVars)
-			config.customEnvVars = message.customEnvVars as Record<string, string>;
-		if (message.customModel) config.customModel = message.customModel as string;
-		if (message.customEffort) config.customEffort = message.customEffort as string;
-		if (message.customContextWindow)
-			config.customContextWindow = message.customContextWindow as number;
-		if (message.customProviderPath)
-			config.customProviderPath = message.customProviderPath as string;
-		if (message.sessionSshRemoteConfig) {
-			config.sessionSshRemoteConfig =
-				message.sessionSshRemoteConfig as CreateSessionConfig['sessionSshRemoteConfig'];
-		}
-		// autoRunFolderPath can be set outside of the agent's cwd (no confinement needed)
-		if (message.autoRunFolderPath) config.autoRunFolderPath = message.autoRunFolderPath as string;
-		const hasConfig = Object.keys(config).length > 0;
-
-		this.callbacks
-			.createSession(name, toolType, cwd, groupId, hasConfig ? config : undefined)
-			.then((result) => {
-				this.send(client, {
-					type: 'create_session_result',
-					success: !!result,
-					sessionId: result?.sessionId,
-					requestId: message.requestId,
-				});
-			})
-			.catch((error) => {
-				this.sendError(client, `Failed to create session: ${error.message}`);
-			});
-	}
-
-	/**
-	 * Handle create_worktree_session message - create a new agent in a git
-	 * worktree branched off an existing parent agent, without an Auto Run
-	 * playbook. The desktop creates the worktree, builds a child session linked
-	 * to the parent, and returns the new agent's session id.
-	 */
-	private handleCreateWorktreeSession(client: WebClient, message: WebClientMessage): void {
-		const parentSessionId = message.parentSessionId as string;
-		const branchName = message.branchName as string;
-
-		if (!parentSessionId || typeof parentSessionId !== 'string') {
-			this.sendError(client, 'Missing or invalid parentSessionId');
-			return;
-		}
-
-		if (!branchName || typeof branchName !== 'string' || branchName.trim() === '') {
-			this.sendError(client, 'Missing or invalid branchName');
-			return;
-		}
-
-		if (message.baseBranch !== undefined && typeof message.baseBranch !== 'string') {
-			this.sendError(client, 'baseBranch must be a string');
-			return;
-		}
-
-		if (!this.callbacks.createWorktreeSession) {
-			this.sendError(client, 'Worktree session creation not configured');
-			return;
-		}
-
-		const config = {
-			branchName: branchName.trim(),
-			baseBranch: (message.baseBranch as string | undefined)?.trim() || undefined,
-		};
-
-		this.callbacks
-			.createWorktreeSession(parentSessionId, config)
-			.then((result) => {
-				this.send(client, {
-					type: 'create_worktree_session_result',
-					success: result.success,
-					sessionId: result.sessionId,
-					error: result.error,
-					requestId: message.requestId,
-				});
-			})
-			.catch((error) => {
-				this.sendError(client, `Failed to create worktree session: ${error.message}`);
-			});
-	}
-
-	/**
-	 * Handle delete_session message - delete an agent session
-	 */
-	private handleDeleteSession(client: WebClient, message: WebClientMessage): void {
-		const sessionId = message.sessionId as string;
-
-		if (!sessionId) {
-			this.sendError(client, 'Missing sessionId');
-			return;
-		}
-
-		if (!this.callbacks.deleteSession) {
-			this.sendError(client, 'Session deletion not configured');
-			return;
-		}
-
-		this.callbacks
-			.deleteSession(sessionId)
-			.then((success) => {
-				this.send(client, {
-					type: 'delete_session_result',
-					success,
-					sessionId,
-					requestId: message.requestId,
-				});
-			})
-			.catch((error) => {
-				this.sendError(client, `Failed to delete session: ${error.message}`);
-			});
-	}
-
-	/**
-	 * Handle rename_session message - rename an agent session
-	 */
-	private handleRenameSession(client: WebClient, message: WebClientMessage): void {
-		const sessionId = message.sessionId as string;
-		const newName = message.newName as string;
-
-		if (!sessionId) {
-			this.sendError(client, 'Missing sessionId');
-			return;
-		}
-
-		if (!newName || typeof newName !== 'string' || newName.length === 0) {
-			this.sendError(client, 'Missing or empty newName');
-			return;
-		}
-
-		if (newName.length > 100) {
-			this.sendError(client, 'newName must be 100 characters or less');
-			return;
-		}
-
-		if (!this.callbacks.renameSession) {
-			this.sendError(client, 'Session renaming not configured');
-			return;
-		}
-
-		this.callbacks
-			.renameSession(sessionId, newName)
-			.then((success) => {
-				this.send(client, {
-					type: 'rename_session_result',
-					success,
-					sessionId,
-					newName,
-					requestId: message.requestId,
-				});
-			})
-			.catch((error) => {
-				this.sendError(client, `Failed to rename session: ${error.message}`);
-			});
-	}
-
-	/**
-	 * Handle update_session_cwd message - update an agent's working directory.
-	 * The desktop's `projectRoot` (used for provider session storage) is left
-	 * untouched so historical conversations stay addressable; only the UI-facing
-	 * `cwd`/`fullPath` move. Renderer-side validation rejects updates while an
-	 * agent process is alive - the PTY's cwd is fixed at spawn time.
-	 */
-	private handleUpdateSessionCwd(client: WebClient, message: WebClientMessage): void {
-		const sessionId = message.sessionId as string;
-		const newCwd = message.newCwd as string;
-
-		if (!sessionId) {
-			this.sendError(client, 'Missing sessionId');
-			return;
-		}
-
-		if (!newCwd || typeof newCwd !== 'string' || newCwd.trim() === '') {
-			this.sendError(client, 'Missing or empty newCwd');
-			return;
-		}
-
-		if (!this.callbacks.updateSessionCwd) {
-			this.sendError(client, 'Session cwd updates not configured');
-			return;
-		}
-
-		this.callbacks
-			.updateSessionCwd(sessionId, newCwd)
-			.then((result) => {
-				this.send(client, {
-					type: 'update_session_cwd_result',
-					success: result.success,
-					error: result.error,
-					sessionId,
-					newCwd,
-					requestId: message.requestId,
-				});
-			})
-			.catch((error) => {
-				this.sendError(client, `Failed to update session cwd: ${error.message}`);
-			});
-	}
-
-	/**
-	 * Handle update_session_ssh message - update an agent's SSH execution config
-	 * (remote selection, working-dir override, history-sync flags). Only the
-	 * fields present in `sshPatch` are merged onto the existing config. Like
-	 * cwd updates, the renderer refuses while the agent process is alive because
-	 * the spawn target is fixed at launch time.
-	 */
-	private handleUpdateSessionSsh(client: WebClient, message: WebClientMessage): void {
-		const sessionId = message.sessionId as string;
-		const sshPatch = message.sshPatch as Record<string, unknown> | undefined;
-
-		if (!sessionId) {
-			this.sendError(client, 'Missing sessionId');
-			return;
-		}
-
-		if (!sshPatch || typeof sshPatch !== 'object') {
-			this.sendError(client, 'Missing or invalid sshPatch');
-			return;
-		}
-
-		if (!this.callbacks.updateSessionSsh) {
-			this.sendError(client, 'Session SSH updates not configured');
-			return;
-		}
-
-		this.callbacks
-			.updateSessionSsh(sessionId, sshPatch)
-			.then((result) => {
-				this.send(client, {
-					type: 'update_session_ssh_result',
-					success: result.success,
-					error: result.error,
-					sessionId,
-					requestId: message.requestId,
-				});
-			})
-			.catch((error) => {
-				this.sendError(client, `Failed to update session SSH config: ${error.message}`);
-			});
-	}
-
-	/**
-	 * Handle update_session_config message - update an agent's editable
-	 * per-session config (nudge / new-session message, custom path / args / env
-	 * vars, model, effort, context window, Claude token-source tri-state). Only
-	 * the keys present in `configPatch` are applied; a key with value `null`
-	 * clears that field. These are spawn-time settings (they take effect on the
-	 * next launch), so unlike cwd/SSH the renderer applies them even while the
-	 * agent process is alive.
-	 */
-	private handleUpdateSessionConfig(client: WebClient, message: WebClientMessage): void {
-		const sessionId = message.sessionId as string;
-		const configPatch = message.configPatch as Record<string, unknown> | undefined;
-
-		if (!sessionId) {
-			this.sendError(client, 'Missing sessionId');
-			return;
-		}
-
-		if (!configPatch || typeof configPatch !== 'object') {
-			this.sendError(client, 'Missing or invalid configPatch');
-			return;
-		}
-
-		if (!this.callbacks.updateSessionConfig) {
-			this.sendError(client, 'Session config updates not configured');
-			return;
-		}
-
-		this.callbacks
-			.updateSessionConfig(sessionId, configPatch)
-			.then((result) => {
-				this.send(client, {
-					type: 'update_session_config_result',
-					success: result.success,
-					error: result.error,
-					sessionId,
-					requestId: message.requestId,
-				});
-			})
-			.catch((error) => {
-				this.sendError(client, `Failed to update session config: ${error.message}`);
-			});
-	}
-
-	/**
-	 * Handle get_groups message - return list of groups
-	 */
-	private handleGetGroups(client: WebClient, message: WebClientMessage): void {
-		if (!this.callbacks.getGroups) {
-			this.sendError(client, 'Groups not configured');
-			return;
-		}
-
-		const groups = this.callbacks.getGroups();
-		this.send(client, {
-			type: 'groups_list',
-			groups,
-			requestId: message.requestId,
-		});
-	}
-
-	/**
-	 * Handle create_group message - create a new group
-	 */
-	private handleCreateGroup(client: WebClient, message: WebClientMessage): void {
-		const name = message.name as string;
-		const emoji = message.emoji as string | undefined;
-		const requestedParentGroupId = message.parentGroupId;
-
-		if (
-			requestedParentGroupId !== undefined &&
-			(typeof requestedParentGroupId !== 'string' || !requestedParentGroupId.trim())
-		) {
-			this.sendError(client, 'Invalid parentGroupId');
-			return;
-		}
-
-		const parentGroupId = requestedParentGroupId?.trim();
-
-		if (!name || typeof name !== 'string') {
-			this.sendError(client, 'Missing or invalid group name');
-			return;
-		}
-
-		if (!this.callbacks.createGroup) {
-			this.sendError(client, 'Group creation not configured');
-			return;
-		}
-
-		this.callbacks
-			.createGroup(name, emoji, parentGroupId)
-			.then((result) => {
-				this.send(client, {
-					type: 'create_group_result',
-					success: !!result,
-					groupId: result?.id,
-					requestId: message.requestId,
-				});
-			})
-			.catch((error) => {
-				this.sendError(client, `Failed to create group: ${error.message}`);
-			});
-	}
-
-	/**
-	 * Handle rename_group message - rename a group
-	 */
-	private handleRenameGroup(client: WebClient, message: WebClientMessage): void {
-		const groupId = message.groupId as string;
-		const name = message.name as string;
-
-		if (!groupId) {
-			this.sendError(client, 'Missing groupId');
-			return;
-		}
-
-		if (!name || typeof name !== 'string') {
-			this.sendError(client, 'Missing or invalid group name');
-			return;
-		}
-
-		if (!this.callbacks.renameGroup) {
-			this.sendError(client, 'Group renaming not configured');
-			return;
-		}
-
-		this.callbacks
-			.renameGroup(groupId, name)
-			.then((success) => {
-				this.send(client, {
-					type: 'rename_group_result',
-					success,
-					groupId,
-					requestId: message.requestId,
-				});
-			})
-			.catch((error) => {
-				this.sendError(client, `Failed to rename group: ${error.message}`);
-			});
-	}
-
-	/**
-	 * Handle delete_group message - delete a group
-	 */
-	private handleDeleteGroup(client: WebClient, message: WebClientMessage): void {
-		const groupId = message.groupId as string;
-
-		if (!groupId) {
-			this.sendError(client, 'Missing groupId');
-			return;
-		}
-
-		if (!this.callbacks.deleteGroup) {
-			this.sendError(client, 'Group deletion not configured');
-			return;
-		}
-
-		this.callbacks
-			.deleteGroup(groupId)
-			.then((success) => {
-				this.send(client, {
-					type: 'delete_group_result',
-					success,
-					groupId,
-					requestId: message.requestId,
-				});
-			})
-			.catch((error) => {
-				this.sendError(client, `Failed to delete group: ${error.message}`);
-			});
-	}
-
-	/**
-	 * Handle move_session_to_group message - move a session to a group (or ungrouped)
-	 */
-	private handleMoveSessionToGroup(client: WebClient, message: WebClientMessage): void {
-		const sessionId = message.sessionId as string;
-		const groupId = message.groupId as string | null;
-
-		if (!sessionId) {
-			this.sendError(client, 'Missing sessionId');
-			return;
-		}
-
-		// groupId can be null (for ungrouped), but must be present in message
-		if (!('groupId' in message)) {
-			this.sendError(client, 'Missing groupId (use null for ungrouped)');
-			return;
-		}
-
-		if (!this.callbacks.moveSessionToGroup) {
-			this.sendError(client, 'Move to group not configured');
-			return;
-		}
-
-		this.callbacks
-			.moveSessionToGroup(sessionId, groupId)
-			.then((success) => {
-				this.send(client, {
-					type: 'move_session_to_group_result',
-					success,
-					sessionId,
-					groupId,
-					requestId: message.requestId,
-				});
-			})
-			.catch((error) => {
-				this.sendError(client, `Failed to move session to group: ${error.message}`);
-			});
-	}
-
-	/**
-	 * Handle get_git_status message - fetch git status for a session
-	 */
-	private handleGetGitStatus(client: WebClient, message: WebClientMessage): void {
-		const sessionId = message.sessionId as string;
-
-		if (!sessionId) {
-			this.sendError(client, 'Missing sessionId');
-			return;
-		}
-
-		if (!this.callbacks.getGitStatus) {
-			this.sendError(client, 'Git status not configured');
-			return;
-		}
-
-		this.callbacks
-			.getGitStatus(sessionId)
-			.then((status) => {
-				this.send(client, {
-					type: 'git_status',
-					sessionId,
-					status,
-					requestId: message.requestId,
-				});
-			})
-			.catch((error) => {
-				this.sendError(client, `Failed to get git status: ${error.message}`);
-			});
-	}
-
-	/**
-	 * Handle get_git_diff message - fetch git diff for a session
-	 */
-	private handleGetGitDiff(client: WebClient, message: WebClientMessage): void {
-		const sessionId = message.sessionId as string;
-		const filePath = message.filePath as string | undefined;
-
-		if (!sessionId) {
-			this.sendError(client, 'Missing sessionId');
-			return;
-		}
-
-		if (!this.callbacks.getGitDiff) {
-			this.sendError(client, 'Git diff not configured');
-			return;
-		}
-
-		this.callbacks
-			.getGitDiff(sessionId, filePath)
-			.then((diff) => {
-				this.send(client, {
-					type: 'git_diff',
-					sessionId,
-					diff,
-					requestId: message.requestId,
-				});
-			})
-			.catch((error) => {
-				this.sendError(client, `Failed to get git diff: ${error.message}`);
-			});
-	}
-
-	/**
-	 * Handle get_git_branches message - list local + remote branches for a session's
-	 * cwd, used by the mobile Run-in-Worktree base-branch picker.
-	 */
-	private handleGetGitBranches(client: WebClient, message: WebClientMessage): void {
-		const sessionId = message.sessionId as string;
-
-		if (!sessionId) {
-			this.sendError(client, 'Missing sessionId');
-			return;
-		}
-
-		if (!this.callbacks.getGitBranchesForSession) {
-			this.sendError(client, 'Git branches not configured');
-			return;
-		}
-
-		this.callbacks
-			.getGitBranchesForSession(sessionId)
-			.then((result) => {
-				this.send(client, {
-					type: 'git_branches',
-					sessionId,
-					branches: result.branches,
-					currentBranch: result.currentBranch,
-					requestId: message.requestId,
-				});
-			})
-			.catch((error: unknown) => {
-				this.reportHandlerError(
-					client,
-					error,
-					'get_git_branches',
-					{ sessionId, requestId: message.requestId },
-					'Failed to get git branches'
-				);
-			});
-	}
-
-	/**
-	 * Handle list_worktrees message - list existing worktrees for a session's cwd,
-	 * used by mobile Run-in-Worktree to offer "use existing" alongside "create new".
-	 */
-	private handleListWorktrees(client: WebClient, message: WebClientMessage): void {
-		const sessionId = message.sessionId as string;
-
-		if (!sessionId) {
-			this.sendError(client, 'Missing sessionId');
-			return;
-		}
-
-		if (!this.callbacks.listWorktreesForSession) {
-			this.sendError(client, 'List worktrees not configured');
-			return;
-		}
-
-		this.callbacks
-			.listWorktreesForSession(sessionId)
-			.then((result) => {
-				this.send(client, {
-					type: 'worktrees_list',
-					sessionId,
-					worktrees: result.worktrees,
-					requestId: message.requestId,
-				});
-			})
-			.catch((error: unknown) => {
-				this.reportHandlerError(
-					client,
-					error,
-					'list_worktrees',
-					{ sessionId, requestId: message.requestId },
-					'Failed to list worktrees'
-				);
-			});
-	}
-
-	/**
-	 * Handle get_group_chats message - return list of all group chats
-	 */
-	private handleGetGroupChats(client: WebClient, message: WebClientMessage): void {
-		if (!this.callbacks.getGroupChats) {
-			this.sendError(client, 'Group chats not configured');
-			return;
-		}
-
-		this.callbacks
-			.getGroupChats()
-			.then((chats) => {
-				this.send(client, {
-					type: 'group_chats_list',
-					chats,
-					requestId: message.requestId,
-				});
-			})
-			.catch((error) => {
-				this.sendError(client, `Failed to get group chats: ${error.message}`);
-			});
-	}
-
-	/**
-	 * Handle start_group_chat message - start a new group chat
-	 */
-	private handleStartGroupChat(client: WebClient, message: WebClientMessage): void {
-		const topic = message.topic as string;
-		const participantIds = message.participantIds as string[];
-
-		if (!topic || typeof topic !== 'string') {
-			this.sendError(client, 'Missing or invalid topic');
-			return;
-		}
-
-		if (!participantIds || !Array.isArray(participantIds) || participantIds.length < 2) {
-			this.sendError(client, 'At least 2 participants are required');
-			return;
-		}
-
-		if (!this.callbacks.startGroupChat) {
-			this.sendError(client, 'Group chat not configured');
-			return;
-		}
-
-		this.callbacks
-			.startGroupChat(topic, participantIds)
-			.then((result) => {
-				this.send(client, {
-					type: 'start_group_chat_result',
-					success: !!result,
-					chatId: result?.chatId,
-					requestId: message.requestId,
-				});
-			})
-			.catch((error) => {
-				this.sendError(client, `Failed to start group chat: ${error.message}`);
-			});
-	}
-
-	/**
-	 * Handle get_group_chat_state message - get state of a specific group chat
-	 */
-	private handleGetGroupChatState(client: WebClient, message: WebClientMessage): void {
-		const chatId = message.chatId as string;
-
-		if (!chatId) {
-			this.sendError(client, 'Missing chatId');
-			return;
-		}
-
-		if (!this.callbacks.getGroupChatState) {
-			this.sendError(client, 'Group chat not configured');
-			return;
-		}
-
-		this.callbacks
-			.getGroupChatState(chatId)
-			.then((state) => {
-				this.send(client, {
-					type: 'group_chat_state',
-					chatId,
-					state,
-					requestId: message.requestId,
-				});
-			})
-			.catch((error) => {
-				this.sendError(client, `Failed to get group chat state: ${error.message}`);
-			});
-	}
-
-	/**
-	 * Handle send_group_chat_message message - send a message to a group chat
-	 */
-	private handleSendGroupChatMessage(client: WebClient, message: WebClientMessage): void {
-		const chatId = message.chatId as string;
-		const chatMessage = message.message as string;
-
-		if (!chatId) {
-			this.sendError(client, 'Missing chatId');
-			return;
-		}
-
-		if (!chatMessage || typeof chatMessage !== 'string') {
-			this.sendError(client, 'Missing or invalid message');
-			return;
-		}
-
-		if (!this.callbacks.sendGroupChatMessage) {
-			this.sendError(client, 'Group chat not configured');
-			return;
-		}
-
-		this.callbacks
-			.sendGroupChatMessage(chatId, chatMessage)
-			.then((success) => {
-				this.send(client, {
-					type: 'send_group_chat_message_result',
-					success,
-					chatId,
-					requestId: message.requestId,
-				});
-			})
-			.catch((error) => {
-				this.sendError(client, `Failed to send group chat message: ${error.message}`);
-			});
-	}
-
-	/**
-	 * Handle stop_group_chat message - stop an active group chat
-	 */
-	private handleStopGroupChat(client: WebClient, message: WebClientMessage): void {
-		const chatId = message.chatId as string;
-
-		if (!chatId) {
-			this.sendError(client, 'Missing chatId');
-			return;
-		}
-
-		if (!this.callbacks.stopGroupChat) {
-			this.sendError(client, 'Group chat not configured');
-			return;
-		}
-
-		this.callbacks
-			.stopGroupChat(chatId)
-			.then((success) => {
-				this.send(client, {
-					type: 'stop_group_chat_result',
-					success,
-					chatId,
-					requestId: message.requestId,
-				});
-			})
-			.catch((error) => {
-				this.sendError(client, `Failed to stop group chat: ${error.message}`);
-			});
-	}
-
-	/**
-	 * Handle merge_context message - merge context from source to target session
-	 */
-	private handleMergeContext(client: WebClient, message: WebClientMessage): void {
-		const sourceSessionId = message.sourceSessionId as string;
-		const targetSessionId = message.targetSessionId as string;
-
-		if (!sourceSessionId || !targetSessionId) {
-			this.sendError(client, 'Missing sourceSessionId or targetSessionId');
-			return;
-		}
-
-		if (sourceSessionId === targetSessionId) {
-			this.sendError(client, 'Source and target sessions must be different');
-			return;
-		}
-
-		if (!this.callbacks.mergeContext) {
-			this.sendError(client, 'Context merge not configured');
-			return;
-		}
-
-		this.callbacks
-			.mergeContext(sourceSessionId, targetSessionId)
-			.then((success) => {
-				this.send(client, {
-					type: 'merge_context_result',
-					success,
-					requestId: message.requestId,
-					timestamp: Date.now(),
-				});
-			})
-			.catch((error) => {
-				this.sendError(client, `Failed to merge context: ${error.message}`);
-			});
-	}
-
-	/**
-	 * Handle transfer_context message - transfer context from source to target session
-	 */
-	private handleTransferContext(client: WebClient, message: WebClientMessage): void {
-		const sourceSessionId = message.sourceSessionId as string;
-		const targetSessionId = message.targetSessionId as string;
-
-		if (!sourceSessionId || !targetSessionId) {
-			this.sendError(client, 'Missing sourceSessionId or targetSessionId');
-			return;
-		}
-
-		if (sourceSessionId === targetSessionId) {
-			this.sendError(client, 'Source and target sessions must be different');
-			return;
-		}
-
-		if (!this.callbacks.transferContext) {
-			this.sendError(client, 'Context transfer not configured');
-			return;
-		}
-
-		this.callbacks
-			.transferContext(sourceSessionId, targetSessionId)
-			.then((success) => {
-				this.send(client, {
-					type: 'transfer_context_result',
-					success,
-					requestId: message.requestId,
-					timestamp: Date.now(),
-				});
-			})
-			.catch((error) => {
-				this.sendError(client, `Failed to transfer context: ${error.message}`);
-			});
-	}
-
-	/**
-	 * Handle summarize_context message - summarize context for a session
-	 */
-	private handleSummarizeContext(client: WebClient, message: WebClientMessage): void {
-		const sessionId = message.sessionId as string;
-
-		if (!sessionId) {
-			this.sendError(client, 'Missing sessionId');
-			return;
-		}
-
-		if (!this.callbacks.summarizeContext) {
-			this.sendError(client, 'Context summarize not configured');
-			return;
-		}
-
-		this.callbacks
-			.summarizeContext(sessionId)
-			.then((success) => {
-				this.send(client, {
-					type: 'summarize_context_result',
-					success,
-					requestId: message.requestId,
-					timestamp: Date.now(),
-				});
-			})
-			.catch((error) => {
-				this.sendError(client, `Failed to summarize context: ${error.message}`);
-			});
-	}
-
-	/**
-	 * Handle create_gist message - publish a session's transcript to a GitHub gist.
-	 * Always replies with `create_gist_result` (even on failure) so waiting
-	 * clients don't hang until their request timeout.
-	 */
-	private handleCreateGist(client: WebClient, message: WebClientMessage): void {
-		const reply = (result: { success: boolean; gistUrl?: string; error?: string }) => {
-			this.send(client, {
-				type: 'create_gist_result',
-				...result,
-				requestId: message.requestId,
-			});
-		};
-
-		const sessionId = message.sessionId;
-		if (typeof sessionId !== 'string' || !sessionId) {
-			reply({ success: false, error: 'Missing sessionId' });
-			return;
-		}
-
-		// Strict validation - avoid truthy coercion so a string like "false"
-		// cannot flip a private gist to public.
-		if (message.description !== undefined && typeof message.description !== 'string') {
-			reply({ success: false, error: 'description must be a string when provided' });
-			return;
-		}
-		if (message.isPublic !== undefined && typeof message.isPublic !== 'boolean') {
-			reply({ success: false, error: 'isPublic must be a boolean when provided' });
-			return;
-		}
-		const description = message.description ?? '';
-		const isPublic = message.isPublic ?? false;
-
-		if (!this.callbacks.createGist) {
-			reply({ success: false, error: 'Gist creation not configured' });
-			return;
-		}
-
-		this.callbacks
-			.createGist(sessionId, description, isPublic)
-			.then((result) => {
-				reply(result);
-			})
-			.catch((error: unknown) => {
-				const msg = error instanceof Error ? error.message : String(error);
-				reply({ success: false, error: `Failed to create gist: ${msg}` });
-			});
-	}
-
-	/**
-	 * Handle get_cue_subscriptions message - fetch Cue subscriptions
-	 */
-	private handleGetCueSubscriptions(client: WebClient, message: WebClientMessage): void {
-		const sessionId = message.sessionId as string | undefined;
-
-		if (!this.callbacks.getCueSubscriptions) {
-			this.sendError(client, 'Cue subscriptions not available');
-			return;
-		}
-
-		this.callbacks
-			.getCueSubscriptions(sessionId)
-			.then((subscriptions) => {
-				this.send(client, {
-					type: 'cue_subscriptions',
-					subscriptions,
-					requestId: message.requestId,
-					timestamp: Date.now(),
-				});
-			})
-			.catch((error) => {
-				this.sendError(client, `Failed to get Cue subscriptions: ${error.message}`);
-			});
-	}
-
-	/**
-	 * Handle toggle_cue_subscription message - enable/disable a subscription
-	 */
-	private handleToggleCueSubscription(client: WebClient, message: WebClientMessage): void {
-		const subscriptionId = message.subscriptionId as string;
-		const enabled = message.enabled as boolean;
-
-		if (!subscriptionId) {
-			this.sendError(client, 'Missing subscriptionId');
-			return;
-		}
-
-		if (typeof enabled !== 'boolean') {
-			this.sendError(client, 'Missing or invalid enabled flag');
-			return;
-		}
-
-		if (!this.callbacks.toggleCueSubscription) {
-			this.sendError(client, 'Cue toggle not available');
-			return;
-		}
-
-		this.callbacks
-			.toggleCueSubscription(subscriptionId, enabled)
-			.then((success) => {
-				this.send(client, {
-					type: 'toggle_cue_subscription_result',
-					success,
-					subscriptionId,
-					enabled,
-					requestId: message.requestId,
-					timestamp: Date.now(),
-				});
-			})
-			.catch((error) => {
-				this.sendError(client, `Failed to toggle Cue subscription: ${error.message}`);
-			});
-	}
-
-	/**
-	 * Handle get_cue_activity message - fetch Cue activity log
-	 */
-	private handleGetCueActivity(client: WebClient, message: WebClientMessage): void {
-		const sessionId = message.sessionId as string | undefined;
-		const limit = (message.limit as number) ?? 50;
-
-		if (!this.callbacks.getCueActivity) {
-			this.sendError(client, 'Cue activity not available');
-			return;
-		}
-
-		this.callbacks
-			.getCueActivity(sessionId, limit)
-			.then((entries) => {
-				this.send(client, {
-					type: 'cue_activity',
-					entries,
-					requestId: message.requestId,
-					timestamp: Date.now(),
-				});
-			})
-			.catch((error) => {
-				this.sendError(client, `Failed to get Cue activity: ${error.message}`);
-			});
-	}
-
-	/**
-	 * Handle trigger_cue_subscription message - manually trigger a Cue subscription
-	 */
-	private handleTriggerCueSubscription(client: WebClient, message: WebClientMessage): void {
-		const subscriptionName = message.subscriptionName;
-		const prompt = message.prompt;
-
-		if (typeof subscriptionName !== 'string' || subscriptionName.trim() === '') {
-			this.sendError(client, 'Missing subscriptionName');
-			return;
-		}
-		if (prompt !== undefined && typeof prompt !== 'string') {
-			this.sendError(client, 'Invalid prompt: must be a string when provided');
-			return;
-		}
-
-		if (!this.callbacks.triggerCueSubscription) {
-			this.sendError(client, 'Cue trigger not available');
-			return;
-		}
-
-		const rawSourceAgentId = message.sourceAgentId;
-		if (rawSourceAgentId !== undefined && typeof rawSourceAgentId !== 'string') {
-			this.sendError(client, 'Invalid sourceAgentId: must be a string when provided');
-			return;
-		}
-		const sourceAgentId = rawSourceAgentId as string | undefined;
-
-		this.callbacks
-			.triggerCueSubscription(subscriptionName, prompt as string | undefined, sourceAgentId)
-			.then((success) => {
-				this.send(client, {
-					type: 'trigger_cue_subscription_result',
-					success,
-					subscriptionName,
-					requestId: message.requestId,
-					timestamp: Date.now(),
-				});
-			})
-			.catch((error) => {
-				const err = error instanceof Error ? error : new Error(String(error));
-				logger.error(`Failed to trigger Cue subscription: ${err.message}`, 'WebSocket');
-				this.sendError(client, `Failed to trigger Cue subscription: ${err.message}`);
-			});
-	}
-
-	/**
-	 * Handle cue_pipeline_list - return all named pipeline entries from
-	 * the on-disk cue-pipeline-layout.json. Pipelines are returned as
-	 * opaque JSON objects so the CLI doesn't need to share the editor's
-	 * full type tree to round-trip them.
-	 */
-	private handleCuePipelineList(client: WebClient, message: WebClientMessage): void {
-		if (!this.callbacks.listCuePipelines) {
-			this.sendError(client, 'Cue pipeline list not available');
-			return;
-		}
-		this.callbacks
-			.listCuePipelines()
-			.then(({ pipelines }) => {
-				this.send(client, {
-					type: 'cue_pipeline_list_result',
-					pipelines,
-					requestId: message.requestId,
-					timestamp: Date.now(),
-				});
-			})
-			.catch((error) => {
-				const err = error instanceof Error ? error : new Error(String(error));
-				logger.error(`Failed to list Cue pipelines: ${err.message}`, 'WebSocket');
-				this.sendError(client, `Failed to list Cue pipelines: ${err.message}`);
-			});
-	}
-
-	/**
-	 * Handle cue_pipeline_get - fetch a single pipeline entry by name or
-	 * id. Missing entries respond with `pipeline: null` rather than an
-	 * error so scripts can treat "not found" as a normal value.
-	 */
-	private handleCuePipelineGet(client: WebClient, message: WebClientMessage): void {
-		const identifier = message.identifier;
-		if (typeof identifier !== 'string' || identifier.length === 0) {
-			this.sendError(client, 'Missing identifier (pipeline name or id)');
-			return;
-		}
-		if (!this.callbacks.getCuePipeline) {
-			this.sendError(client, 'Cue pipeline get not available');
-			return;
-		}
-
-		this.callbacks
-			.getCuePipeline(identifier)
-			.then((pipeline) => {
-				this.send(client, {
-					type: 'cue_pipeline_get_result',
-					pipeline,
-					requestId: message.requestId,
-					timestamp: Date.now(),
-				});
-			})
-			.catch((error) => {
-				const err = error instanceof Error ? error : new Error(String(error));
-				logger.error(`Failed to get Cue pipeline: ${err.message}`, 'WebSocket');
-				this.sendError(client, `Failed to get Cue pipeline: ${err.message}`);
-			});
-	}
-
-	/**
-	 * Handle cue_pipeline_set - add or replace a pipeline entry. The
-	 * callback returns a structured result so the CLI can map error codes
-	 * (already_exists / not_found / invalid_input / …) to non-zero exit
-	 * codes without parsing free-form messages.
-	 */
-	private handleCuePipelineSet(client: WebClient, message: WebClientMessage): void {
-		const identifier = message.identifier;
-		const pipeline = message.pipeline;
-		const policyRaw = message.policy;
-		if (typeof identifier !== 'string' || identifier.length === 0) {
-			this.sendError(client, 'Missing identifier (pipeline name or id)');
-			return;
-		}
-		if (policyRaw !== 'add' && policyRaw !== 'replace') {
-			this.sendError(client, 'Invalid policy: must be "add" or "replace"');
-			return;
-		}
-		if (pipeline === undefined || pipeline === null) {
-			this.sendError(client, 'Missing pipeline payload');
-			return;
-		}
-		if (!this.callbacks.setCuePipeline) {
-			this.sendError(client, 'Cue pipeline set not available');
-			return;
-		}
-
-		this.callbacks
-			.setCuePipeline(identifier, pipeline, policyRaw)
-			.then((result) => {
-				this.send(client, {
-					type: 'cue_pipeline_set_result',
-					result,
-					requestId: message.requestId,
-					timestamp: Date.now(),
-				});
-			})
-			.catch((error) => {
-				const err = error instanceof Error ? error : new Error(String(error));
-				logger.error(`Failed to set Cue pipeline: ${err.message}`, 'WebSocket');
-				this.sendError(client, `Failed to set Cue pipeline: ${err.message}`);
-			});
-	}
-
-	/**
-	 * Handle cue_pipeline_remove - delete a pipeline entry by name or id.
-	 */
-	private handleCuePipelineRemove(client: WebClient, message: WebClientMessage): void {
-		const identifier = message.identifier;
-		if (typeof identifier !== 'string' || identifier.length === 0) {
-			this.sendError(client, 'Missing identifier (pipeline name or id)');
-			return;
-		}
-		if (!this.callbacks.removeCuePipeline) {
-			this.sendError(client, 'Cue pipeline remove not available');
-			return;
-		}
-
-		this.callbacks
-			.removeCuePipeline(identifier)
-			.then((result) => {
-				this.send(client, {
-					type: 'cue_pipeline_remove_result',
-					result,
-					requestId: message.requestId,
-					timestamp: Date.now(),
-				});
-			})
-			.catch((error) => {
-				const err = error instanceof Error ? error : new Error(String(error));
-				logger.error(`Failed to remove Cue pipeline: ${err.message}`, 'WebSocket');
-				this.sendError(client, `Failed to remove Cue pipeline: ${err.message}`);
-			});
-	}
-
-	/**
-	 * Handle get_usage_dashboard message - fetch usage analytics data
-	 */
-	private handleGetUsageDashboard(client: WebClient, message: WebClientMessage): void {
-		const timeRange = (message.timeRange as string) || 'week';
-		const validRanges = new Set(['day', 'week', 'month', 'all']);
-
-		if (!validRanges.has(timeRange)) {
-			this.sendError(client, 'Invalid timeRange. Must be one of: day, week, month, all');
-			return;
-		}
-
-		if (!this.callbacks.getUsageDashboard) {
-			this.sendError(client, 'Usage dashboard not available');
-			return;
-		}
-
-		this.callbacks
-			.getUsageDashboard(timeRange as 'day' | 'week' | 'month' | 'all')
-			.then((data) => {
-				this.send(client, {
-					type: 'usage_dashboard',
-					data,
-					requestId: message.requestId,
-					timestamp: Date.now(),
-				});
-			})
-			.catch((error) => {
-				this.sendError(client, `Failed to get usage dashboard: ${error.message}`);
-			});
-	}
-
-	/**
-	 * Handle get_stats_aggregation message - return the Usage Dashboard's
-	 * aggregated stats (query counts, durations, per-agent/day/hour breakdowns)
-	 * for a time range. Reads the main-process stats singleton directly.
-	 */
-	private handleGetStatsAggregation(client: WebClient, message: WebClientMessage): void {
-		const range = (message.range as string) || 'week';
-		const validRanges = new Set(['day', 'week', 'month', 'quarter', 'year', 'all']);
-
-		if (!validRanges.has(range)) {
-			this.sendError(client, 'Invalid range. Must be one of: day, week, month, quarter, year, all');
-			return;
-		}
-
-		try {
-			const data = getStatsDB().getAggregatedStats(range as StatsTimeRange);
-			this.send(client, {
-				type: 'stats_aggregation',
-				data,
-				requestId: message.requestId,
-				timestamp: Date.now(),
-			});
-		} catch (error) {
-			this.sendError(
-				client,
-				`Failed to get stats aggregation: ${error instanceof Error ? error.message : String(error)}`
-			);
-		}
-	}
-
-	/**
-	 * Handle stats_query message - run a single read-only SQL statement against
-	 * the stats database and return the rows. Read-only enforcement lives in
-	 * runReadonlyStatsQuery (dedicated readonly connection + single-statement +
-	 * stmt.readonly assertion).
-	 */
-	private handleStatsQuery(client: WebClient, message: WebClientMessage): void {
-		const sql = message.sql as string | undefined;
-		const params = Array.isArray(message.params) ? (message.params as unknown[]) : [];
-
-		if (!sql || typeof sql !== 'string') {
-			this.sendError(client, 'Missing required "sql" string for stats_query');
-			return;
-		}
-
-		try {
-			const result = runReadonlyStatsQuery(sql, params);
-			this.send(client, {
-				type: 'stats_query_result',
-				columns: result.columns,
-				rows: result.rows,
-				rowCount: result.rowCount,
-				truncated: result.truncated,
-				requestId: message.requestId,
-				timestamp: Date.now(),
-			});
-		} catch (error) {
-			this.sendError(
-				client,
-				`Stats query failed: ${error instanceof Error ? error.message : String(error)}`
-			);
-		}
-	}
-
-	/**
-	 * Handle get_achievements message - fetch achievement data
-	 */
-	private handleGetAchievements(client: WebClient, message: WebClientMessage): void {
-		if (!this.callbacks.getAchievements) {
-			this.sendError(client, 'Achievements not available');
-			return;
-		}
-
-		this.callbacks
-			.getAchievements()
-			.then((achievements) => {
-				this.send(client, {
-					type: 'achievements',
-					achievements,
-					requestId: message.requestId,
-					timestamp: Date.now(),
-				});
-			})
-			.catch((error) => {
-				this.sendError(client, `Failed to get achievements: ${error.message}`);
-			});
-	}
-
-	/**
-	 * Handle generate_director_notes_synopsis - generate AI synopsis via batch-mode agent
-	 */
-	private handleGenerateDirectorNotesSynopsis(client: WebClient, message: WebClientMessage): void {
-		if (!this.callbacks.generateDirectorNotesSynopsis) {
-			this.send(client, {
-				type: 'generate_director_notes_synopsis_result',
-				success: false,
-				error: "Director's Notes synopsis generation not available",
-				requestId: message.requestId,
-			});
-			return;
-		}
-
-		const lookbackDays = (message.lookbackDays as number) || 7;
-		const provider = (message.provider as string) || 'claude-code';
-
-		this.callbacks
-			.generateDirectorNotesSynopsis(lookbackDays, provider)
-			.then((result) => {
-				this.send(client, {
-					type: 'generate_director_notes_synopsis_result',
-					...result,
-					requestId: message.requestId,
-					timestamp: Date.now(),
-				});
-			})
-			.catch((error) => {
-				this.send(client, {
-					type: 'generate_director_notes_synopsis_result',
-					success: false,
-					error: `Synopsis generation failed: ${error.message}`,
-					requestId: message.requestId,
-				});
-			});
-	}
-
-	/**
-	 * Handle terminal_write - write raw data to the terminal PTY
-	 */
-	private handleTerminalWrite(client: WebClient, message: WebClientMessage): void {
-		const sessionId = message.sessionId;
-		const data = message.data as string | undefined;
-		if (!sessionId || typeof data !== 'string') {
-			this.send(client, {
-				type: 'terminal_write_result',
-				success: false,
-				error: 'Missing sessionId or data',
-			});
-			return;
-		}
-		if (client.subscribedSessionId !== sessionId) {
-			this.send(client, {
-				type: 'terminal_write_result',
-				success: false,
-				error: 'Not subscribed to this session',
-			});
-			return;
-		}
-		if (!this.callbacks.writeToTerminal) {
-			this.send(client, {
-				type: 'terminal_write_result',
-				success: false,
-				error: 'writeToTerminal not available',
-			});
-			return;
-		}
-		const success = this.callbacks.writeToTerminal(sessionId, data);
-		this.send(client, { type: 'terminal_write_result', success, sessionId });
-	}
-
-	/**
-	 * Handle terminal_resize - resize the terminal PTY
-	 */
-	private handleTerminalResize(client: WebClient, message: WebClientMessage): void {
-		const sessionId = message.sessionId;
-		const cols = message.cols as number | undefined;
-		const rows = message.rows as number | undefined;
-		if (!sessionId || typeof cols !== 'number' || typeof rows !== 'number') {
-			this.send(client, {
-				type: 'terminal_resize_result',
-				success: false,
-				error: 'Missing sessionId, cols, or rows',
-			});
-			return;
-		}
-		if (client.subscribedSessionId !== sessionId) {
-			this.send(client, {
-				type: 'terminal_resize_result',
-				success: false,
-				error: 'Not subscribed to this session',
-			});
-			return;
-		}
-		if (!this.callbacks.resizeTerminal) {
-			this.send(client, {
-				type: 'terminal_resize_result',
-				success: false,
-				error: 'resizeTerminal not available',
-			});
-			return;
-		}
-		const success = this.callbacks.resizeTerminal(sessionId, cols, rows);
-		this.send(client, { type: 'terminal_resize_result', success, sessionId });
-	}
-
-	/**
-	 * Handle notify_toast - show a toast notification in the desktop app.
-	 */
-	private handleNotifyToast(client: WebClient, message: WebClientMessage): void {
-		const title = typeof message.title === 'string' ? message.title : '';
-		const body = typeof message.message === 'string' ? message.message : '';
-		const rawColor = typeof message.color === 'string' ? message.color : undefined;
-		// Legacy field (kept for back-compat with older CLI scripts).
-		const rawType = typeof message.toastType === 'string' ? message.toastType : undefined;
-		const duration = typeof message.duration === 'number' ? message.duration : undefined;
-		const dismissible = message.dismissible === true;
-		const sessionId = typeof message.sessionId === 'string' ? message.sessionId : undefined;
-		const sourceAgent =
-			typeof message.sourceAgent === 'string' && message.sourceAgent.length > 0
-				? message.sourceAgent
-				: undefined;
-		const tabId = typeof message.tabId === 'string' ? message.tabId : undefined;
-		const actionUrl = typeof message.actionUrl === 'string' ? message.actionUrl : undefined;
-		const actionLabel = typeof message.actionLabel === 'string' ? message.actionLabel : undefined;
-		const rawClickAction =
-			typeof message.clickAction === 'object' && message.clickAction !== null
-				? (message.clickAction as Record<string, unknown>)
-				: undefined;
-
-		const sendResult = (success: boolean, error?: string) => {
-			this.send(client, {
-				type: 'notify_toast_result',
-				success,
-				error,
-				requestId: message.requestId,
-			});
-		};
-
-		if (!title) {
-			sendResult(false, 'Missing title');
-			return;
-		}
-
-		// Resolve color: explicit `color` wins over deprecated `toastType`. Default `theme`.
-		let color: NotifyToastColor;
-		if (rawColor !== undefined) {
-			if (!NOTIFY_TOAST_COLORS.includes(rawColor as NotifyToastColor)) {
-				sendResult(
-					false,
-					`Invalid toast color: ${rawColor}. Must be one of: ${NOTIFY_TOAST_COLORS.join(', ')}`
-				);
-				return;
-			}
-			color = rawColor as NotifyToastColor;
-		} else if (rawType !== undefined) {
-			if (!NOTIFY_TOAST_KINDS.includes(rawType as NotifyToastKind)) {
-				sendResult(false, `Invalid toast type: ${rawType}`);
-				return;
-			}
-			color = VARIANT_TO_COLOR[rawType as NotifyCenterFlashVariant];
-		} else {
-			color = 'theme';
-		}
-
-		// Validate clickAction (data-driven click intent). Each kind has its
-		// own required fields; bad shapes are rejected so the CLI surfaces a
-		// clear error instead of producing a silent no-op toast.
-		let clickAction: NotifyToastClickAction | undefined;
-		if (rawClickAction !== undefined) {
-			const kind = rawClickAction.kind;
-			if (kind === 'jump-session') {
-				const id = rawClickAction.sessionId;
-				if (typeof id !== 'string' || id.length === 0) {
-					sendResult(false, "clickAction kind 'jump-session' requires sessionId");
-					return;
-				}
-				const tab = rawClickAction.tabId;
-				clickAction = {
-					kind: 'jump-session',
-					sessionId: id,
-					tabId: typeof tab === 'string' && tab.length > 0 ? tab : undefined,
-				};
-			} else if (kind === 'open-file') {
-				const id = rawClickAction.sessionId;
-				const path = rawClickAction.path;
-				if (typeof id !== 'string' || id.length === 0) {
-					sendResult(false, "clickAction kind 'open-file' requires sessionId");
-					return;
-				}
-				if (typeof path !== 'string' || path.length === 0) {
-					sendResult(false, "clickAction kind 'open-file' requires path");
-					return;
-				}
-				clickAction = { kind: 'open-file', sessionId: id, path };
-			} else if (kind === 'open-url') {
-				const url = rawClickAction.url;
-				if (typeof url !== 'string' || url.length === 0) {
-					sendResult(false, "clickAction kind 'open-url' requires url");
-					return;
-				}
-				clickAction = { kind: 'open-url', url };
-			} else {
-				sendResult(
-					false,
-					`Invalid clickAction kind: ${String(kind)}. Must be one of: jump-session, open-file, open-url`
-				);
-				return;
-			}
-		}
-
-		// Duration validation: reject 0 (use --dismissible instead) and cap at 60 s.
-		// Skipped entirely when `dismissible: true` (the toast is sticky).
-		if (!dismissible && duration !== undefined) {
-			if (!Number.isFinite(duration) || duration <= 0) {
-				sendResult(
-					false,
-					'duration must be a positive number of seconds (use dismissible:true for sticky toasts)'
-				);
-				return;
-			}
-			if (duration > EXTERNAL_TOAST_MAX_DURATION_SECONDS) {
-				sendResult(
-					false,
-					`duration cannot exceed ${EXTERNAL_TOAST_MAX_DURATION_SECONDS} seconds for externally-triggered toasts (use dismissible:true to make it sticky)`
-				);
-				return;
-			}
-		}
-
-		if (!this.callbacks.notifyToast) {
-			sendResult(false, 'Toast notifications not configured');
-			return;
-		}
-
-		this.callbacks
-			.notifyToast({
-				title,
-				message: body,
-				color,
-				dismissible,
-				duration,
-				sessionId,
-				sourceAgent,
-				tabId,
-				actionUrl,
-				actionLabel,
-				clickAction,
-			})
-			.then((success) => sendResult(success, success ? undefined : 'Failed to show toast'))
-			.catch((error) => sendResult(false, `Failed to show toast: ${error.message}`));
-	}
-
-	/**
-	 * Handle movement - begin/add/update/move/remove/clear an item on the agent-driven
-	 * movement. Validates op + id, then hands a typed payload to the renderer via
-	 * the movementView callback (the `remote:movement` channel).
-	 */
-	private handleMovement(client: WebClient, message: WebClientMessage): void {
-		const op = typeof message.op === 'string' ? (message.op as MovementOp) : undefined;
-		const rawId = typeof message.id === 'string' ? message.id : '';
-		const id = rawId.trim();
-
-		const sendResult = (success: boolean, error?: string) => {
-			this.send(client, { type: 'movement_result', success, error, requestId: message.requestId });
-		};
-
-		if (!op || !MOVEMENT_OPS.includes(op)) {
-			sendResult(false, `Invalid or missing op. Must be one of: ${MOVEMENT_OPS.join(', ')}`);
-			return;
-		}
-		if (op !== 'clear' && !id) {
-			sendResult(false, `Missing movement item id for op '${op}'`);
-			return;
-		}
-		if (op !== 'clear' && rawId !== id) {
-			sendResult(false, 'Movement item id must not contain surrounding whitespace');
-			return;
-		}
-
-		let viewType: MovementViewType | undefined;
-		const rawViewType = typeof message.viewType === 'string' ? message.viewType : undefined;
-		if (rawViewType !== undefined) {
-			if (!MOVEMENT_VIEW_TYPES.includes(rawViewType as MovementViewType)) {
-				sendResult(
-					false,
-					`Invalid viewType: ${rawViewType}. Must be one of: ${MOVEMENT_VIEW_TYPES.join(', ')}`
-				);
-				return;
-			}
-			viewType = rawViewType as MovementViewType;
-		}
-		const body = typeof message.body === 'string' ? message.body : undefined;
-		if ((op === 'add' || op === 'update') && viewType === 'html' && body === undefined) {
-			sendResult(false, `Movement ${op} requires HTML content when viewType is 'html'`);
-			return;
-		}
-		const title = typeof message.title === 'string' ? message.title.trim() : '';
-		if (op === 'begin') {
-			if (!title) {
-				sendResult(false, 'Movement begin requires a non-empty title');
-				return;
-			}
-			if (body !== undefined) {
-				sendResult(false, 'Movement begin uses a host-rendered shell and does not accept content');
-				return;
-			}
-			viewType = 'html';
-		}
-		let phase: ConcertoCreationPhase | undefined;
-		let step: number | undefined;
-		let steps: number | undefined;
-		let notes: ConcertoProgressNote[] | undefined;
-		if (op === 'progress') {
-			const rawPhase = typeof message.phase === 'string' ? message.phase : '';
-			if (!CONCERTO_CREATION_PHASES.includes(rawPhase as ConcertoCreationPhase)) {
-				sendResult(
-					false,
-					`Invalid or missing phase. Must be one of: ${CONCERTO_CREATION_PHASES.join(', ')}`
-				);
-				return;
-			}
-			if (!title) {
-				sendResult(false, 'Movement progress requires a non-empty title');
-				return;
-			}
-			phase = rawPhase as ConcertoCreationPhase;
-			if (message.notes !== undefined) {
-				if (!Array.isArray(message.notes) || message.notes.length === 0) {
-					sendResult(false, 'Invalid notes: expected a non-empty array');
-					return;
-				}
-				notes = [];
-				for (const [index, rawNote] of message.notes.entries()) {
-					if (!rawNote || typeof rawNote !== 'object') {
-						sendResult(false, `Invalid note ${index + 1}: expected an object`);
-						return;
-					}
-					const note = rawNote as Record<string, unknown>;
-					if (!CONCERTO_PROGRESS_NOTE_VALUES.includes(note.value as ConcertoProgressNoteValue)) {
-						sendResult(false, `Invalid note ${index + 1}: unsupported value`);
-						return;
-					}
-					const invalidModifier = (['dotted', 'triad', 'tie'] as const).find(
-						(modifier) => note[modifier] !== undefined && typeof note[modifier] !== 'boolean'
-					);
-					if (invalidModifier) {
-						sendResult(false, `Invalid note ${index + 1}: ${invalidModifier} must be boolean`);
-						return;
-					}
-					notes.push({
-						value: note.value as ConcertoProgressNoteValue,
-						...(note.dotted === true && { dotted: true }),
-						...(note.triad === true && { triad: true }),
-						...(note.tie === true && { tie: true }),
-					});
-				}
-				if (notes.at(-1)?.tie) {
-					sendResult(false, 'Invalid notes: the final note cannot tie forward');
-					return;
-				}
-			}
-			steps = message.steps === undefined ? (notes?.length ?? 1) : Number(message.steps);
-			step = message.step === undefined ? 1 : Number(message.step);
-			if (!Number.isInteger(steps) || steps < 1 || steps > CONCERTO_PROGRESS_MAX_STEPS) {
-				sendResult(
-					false,
-					`Invalid steps: must be an integer from 1 through ${CONCERTO_PROGRESS_MAX_STEPS}`
-				);
-				return;
-			}
-			if (!Number.isInteger(step) || step < 1 || step > steps) {
-				sendResult(false, `Invalid step: must be an integer from 1 through steps (${steps})`);
-				return;
-			}
-			if (notes && notes.length !== steps) {
-				sendResult(false, `Invalid notes: expected exactly ${steps} entries`);
-				return;
-			}
-		}
-
-		// Finite-only: JSON can smuggle Infinity (1e400) or NaN through `typeof
-		// v === 'number'`, which would become invalid CSS geometry downstream.
-		const num = (v: unknown): number | undefined =>
-			typeof v === 'number' && Number.isFinite(v) ? v : undefined;
-		const badGeometry = (['x', 'y', 'width', 'height'] as const).find(
-			(k) => message[k] !== undefined && num(message[k]) === undefined
-		);
-		if (badGeometry) {
-			sendResult(false, `Invalid ${badGeometry}: must be a finite number`);
-			return;
-		}
-		const payload: MovementPayload = {
-			op,
-			id: id || undefined,
-			viewType,
-			x: num(message.x),
-			y: num(message.y),
-			width: num(message.width),
-			height: num(message.height),
-			title: title || undefined,
-			body,
-			phase,
-			step,
-			steps,
-			notes,
-		};
-
-		if (!this.callbacks.movementView) {
-			sendResult(false, 'Movement not configured');
-			return;
-		}
-		this.callbacks
-			.movementView(payload)
-			.then((success) => sendResult(success, success ? undefined : 'Failed to update movement'))
-			.catch((error) => sendResult(false, `Failed to update movement: ${error.message}`));
-	}
-
-	/**
-	 * Handle get_movement_state - return the current movement snapshot (items + size)
-	 * so an agent can compose around what's already placed.
-	 */
-	private handleGetMovementState(client: WebClient, message: WebClientMessage): void {
-		const sendResult = (snapshot: MovementStateSnapshot | null, error?: string) => {
-			this.send(client, {
-				type: 'movement_state_result',
-				success: !error,
-				snapshot,
-				error,
-				requestId: message.requestId,
-			});
-		};
-		if (!this.callbacks.getMovementState) {
-			sendResult(null, 'Movement not configured');
-			return;
-		}
-		this.callbacks
-			.getMovementState()
-			// A null snapshot means the read did NOT happen (flag off, renderer gone,
-			// or timeout) - report failure so callers can retry instead of composing
-			// against a false empty 0x0 layout.
-			.then((snapshot) =>
-				snapshot
-					? sendResult(snapshot)
-					: sendResult(null, 'Movement state unavailable (Concerto off or renderer not responding)')
-			)
-			.catch((error) => sendResult(null, `Failed to read movement state: ${error.message}`));
-	}
-
-	/** Capture a live HTML Movement and return its diagnostics to the CLI. */
-	private handleGetMovementDesignerInspection(client: WebClient, message: WebClientMessage): void {
-		const rawId = typeof message.id === 'string' ? message.id : '';
-		const id = rawId.trim();
-		const sendResult = (inspection: MovementDesignerInspection | null, error?: string) => {
-			this.send(client, {
-				type: 'movement_designer_inspection_result',
-				success: !error,
-				inspection,
-				error,
-				requestId: message.requestId,
-			});
-		};
-		if (!id) {
-			sendResult(null, 'Missing movement item id');
-			return;
-		}
-		if (rawId !== id) {
-			sendResult(null, 'Movement item id must not contain surrounding whitespace');
-			return;
-		}
-		if (!this.callbacks.getMovementDesignerInspection) {
-			sendResult(null, 'Movement designer inspection is not configured');
-			return;
-		}
-		this.callbacks
-			.getMovementDesignerInspection(id)
-			.then((inspection) =>
-				inspection
-					? sendResult(inspection)
-					: sendResult(null, `HTML Movement '${id}' is not visible or did not load`)
-			)
-			.catch((error) => sendResult(null, `Failed to inspect Movement: ${error.message}`));
-	}
-
-	/** Perform a click or text entry inside the sandboxed HTML Movement. */
-	private handleInteractMovementDesigner(client: WebClient, message: WebClientMessage): void {
-		const rawId = typeof message.id === 'string' ? message.id : '';
-		const id = rawId.trim();
-		const rawAction = message.action as Record<string, unknown> | undefined;
-		const selector = typeof rawAction?.selector === 'string' ? rawAction.selector.trim() : '';
-		const kind = rawAction?.kind;
-		const sendResult = (result: ConcertoDesignerActionResult) => {
-			this.send(client, {
-				type: 'movement_designer_interaction_result',
-				success: result.ok,
-				result,
-				error: result.ok ? undefined : result.message,
-				requestId: message.requestId,
-			});
-		};
-		if (!id || !selector || (kind !== 'click' && kind !== 'type')) {
-			sendResult({
-				ok: false,
-				action: kind === 'type' ? 'type' : 'click',
-				selector,
-				message: 'Interaction requires an id, a CSS selector, and click or type action',
-			});
-			return;
-		}
-		if (rawId !== id) {
-			sendResult({
-				ok: false,
-				action: kind,
-				selector,
-				message: 'Movement item id must not contain surrounding whitespace',
-			});
-			return;
-		}
-		if (selector.length > 2048) {
-			sendResult({
-				ok: false,
-				action: kind,
-				selector,
-				message: 'CSS selector is too long',
-			});
-			return;
-		}
-		const value = typeof rawAction?.value === 'string' ? rawAction.value : '';
-		if (kind === 'type' && value.length > 100_000) {
-			sendResult({ ok: false, action: kind, selector, message: 'Input value is too long' });
-			return;
-		}
-		const action: ConcertoDesignerAction =
-			kind === 'click' ? { kind, selector } : { kind, selector, value };
-		if (!this.callbacks.interactMovementDesigner) {
-			sendResult({
-				ok: false,
-				action: action.kind,
-				selector,
-				message: 'Movement designer interaction is not configured',
-			});
-			return;
-		}
-		this.callbacks
-			.interactMovementDesigner(id, action)
-			.then(sendResult)
-			.catch((error) =>
-				sendResult({
-					ok: false,
-					action: action.kind,
-					selector,
-					message: `Failed to interact with Movement: ${error.message}`,
-				})
-			);
-	}
-
-	/**
-	 * Handle cadenza - open/update/close a small agent-driven cadenza view.
-	 * Validates the op/id/viewType/color, then hands a typed payload to the
-	 * renderer via the cadenzaView callback (the `remote:cadenza` channel).
-	 */
-	private handleCadenza(client: WebClient, message: WebClientMessage): void {
-		const op = typeof message.op === 'string' ? (message.op as CadenzaOp) : undefined;
-		const id = typeof message.id === 'string' ? message.id : '';
-
-		const sendResult = (success: boolean, error?: string) => {
-			this.send(client, {
-				type: 'cadenza_result',
-				success,
-				error,
-				requestId: message.requestId,
-			});
-		};
-
-		if (!op || !CADENZA_OPS.includes(op)) {
-			sendResult(false, `Invalid or missing op. Must be one of: ${CADENZA_OPS.join(', ')}`);
-			return;
-		}
-		if (!id) {
-			sendResult(false, 'Missing cadenza id');
-			return;
-		}
-
-		let viewType: CadenzaViewType | undefined;
-		const rawViewType = typeof message.viewType === 'string' ? message.viewType : undefined;
-		if (rawViewType !== undefined) {
-			if (!CADENZA_VIEW_TYPES.includes(rawViewType as CadenzaViewType)) {
-				sendResult(
-					false,
-					`Invalid viewType: ${rawViewType}. Must be one of: ${CADENZA_VIEW_TYPES.join(', ')}`
-				);
-				return;
-			}
-			viewType = rawViewType as CadenzaViewType;
-		}
-		if (op === 'open' && !viewType) {
-			sendResult(false, `op 'open' requires a viewType (${CADENZA_VIEW_TYPES.join(' | ')})`);
-			return;
-		}
-
-		let color: CadenzaColor | undefined;
-		const rawColor = typeof message.color === 'string' ? message.color : undefined;
-		if (rawColor !== undefined) {
-			if (!CADENZA_COLORS.includes(rawColor as CadenzaColor)) {
-				sendResult(
-					false,
-					`Invalid color: ${rawColor}. Must be one of: ${CADENZA_COLORS.join(', ')}`
-				);
-				return;
-			}
-			color = rawColor as CadenzaColor;
-		}
-
-		// Decision buttons: keep only well-formed { label, value } string pairs.
-		const options = Array.isArray(message.options)
-			? (message.options as unknown[]).filter(
-					(o): o is { label: string; value: string } =>
-						!!o &&
-						typeof o === 'object' &&
-						typeof (o as { label?: unknown }).label === 'string' &&
-						typeof (o as { value?: unknown }).value === 'string'
-				)
-			: undefined;
-
-		const payload: CadenzaPayload = {
-			op,
-			id,
-			viewType,
-			title: typeof message.title === 'string' ? message.title : undefined,
-			body: typeof message.body === 'string' ? message.body : undefined,
-			path: typeof message.path === 'string' ? message.path : undefined,
-			options: options && options.length > 0 ? options : undefined,
-			color,
-			sessionId: typeof message.sessionId === 'string' ? message.sessionId : undefined,
-		};
-
-		if (!this.callbacks.cadenzaView) {
-			sendResult(false, 'Cadenza views not configured');
-			return;
-		}
-
-		this.callbacks
-			.cadenzaView(payload)
-			.then((success) => sendResult(success, success ? undefined : 'Failed to update cadenza view'))
-			.catch((error) => sendResult(false, `Failed to update cadenza view: ${error.message}`));
-	}
-
-	/**
-	 * Handle notify_center_flash - show a center-screen flash in the desktop app.
-	 */
-	private handleNotifyCenterFlash(client: WebClient, message: WebClientMessage): void {
-		const body = typeof message.message === 'string' ? message.message : '';
-		const detail = typeof message.detail === 'string' ? message.detail : undefined;
-		const rawColor = typeof message.color === 'string' ? message.color : undefined;
-		const rawVariant = typeof message.variant === 'string' ? message.variant : undefined;
-		const duration = typeof message.duration === 'number' ? message.duration : undefined;
-
-		const sendResult = (success: boolean, error?: string) => {
-			this.send(client, {
-				type: 'notify_center_flash_result',
-				success,
-				error,
-				requestId: message.requestId,
-			});
-		};
-
-		if (!body) {
-			sendResult(false, 'Missing message');
-			return;
-		}
-
-		// Resolve color: explicit `color` wins over deprecated `variant`. Default `theme`.
-		let color: NotifyCenterFlashColor;
-		if (rawColor !== undefined) {
-			if (!NOTIFY_FLASH_COLORS.includes(rawColor as NotifyCenterFlashColor)) {
-				sendResult(
-					false,
-					`Invalid flash color: ${rawColor}. Must be one of: ${NOTIFY_FLASH_COLORS.join(', ')}`
-				);
-				return;
-			}
-			color = rawColor as NotifyCenterFlashColor;
-		} else if (rawVariant !== undefined) {
-			if (!(rawVariant in VARIANT_TO_COLOR)) {
-				sendResult(false, `Invalid flash variant: ${rawVariant}`);
-				return;
-			}
-			color = VARIANT_TO_COLOR[rawVariant as NotifyCenterFlashVariant];
-		} else {
-			color = 'theme';
-		}
-
-		// External flashes must be (0, 5000 ms] - `0` (never auto-dismiss) is rejected so
-		// external scripts can't stick a permanent overlay on the user. In-app callers
-		// using `notifyCenterFlash()` directly are not capped.
-		if (duration !== undefined) {
-			if (!Number.isFinite(duration) || duration <= 0) {
-				sendResult(false, 'duration must be a positive number of milliseconds');
-				return;
-			}
-			if (duration > EXTERNAL_FLASH_MAX_DURATION_MS) {
-				sendResult(
-					false,
-					`duration cannot exceed ${EXTERNAL_FLASH_MAX_DURATION_MS} ms for externally-triggered flashes`
-				);
-				return;
-			}
-		}
-
-		if (!this.callbacks.notifyCenterFlash) {
-			sendResult(false, 'Center flash not configured');
-			return;
-		}
-
-		this.callbacks
-			.notifyCenterFlash({ message: body, detail, color, duration })
-			.then((success) => sendResult(success, success ? undefined : 'Failed to show flash'))
-			.catch((error) => sendResult(false, `Failed to show flash: ${error.message}`));
-	}
-
-	/**
-	 * Handle profiling_start - begin a Chromium contentTracing capture.
-	 *
-	 * contentTracing is an Electron main-process API, so the CLI (a plain bundle
-	 * with no Electron) cannot drive it directly - it routes through here. Mirrors
-	 * the desktop debug:startProfiling handler but with no renderer/dialog coupling
-	 * so it works headlessly for external instrumentation loops.
-	 */
-	private async handleProfilingStart(client: WebClient, message: WebClientMessage): Promise<void> {
-		try {
-			const status = await startProfiling();
-			this.send(client, {
-				type: 'profiling_start_result',
-				success: true,
-				active: status.active,
-				startedAt: status.startedAt,
-				categories: status.categories,
-				requestId: message.requestId,
-			});
-		} catch (error) {
-			const errMsg = error instanceof Error ? error.message : String(error);
-			this.send(client, {
-				type: 'profiling_start_result',
-				success: false,
-				error: `Failed to start profiling: ${errMsg}`,
-				requestId: message.requestId,
-			});
-		}
-	}
-
-	/**
-	 * Handle profiling_status - report whether a capture is in flight.
-	 */
-	private handleProfilingStatus(client: WebClient, message: WebClientMessage): void {
-		const status = getProfilingStatus();
-		this.send(client, {
-			type: 'profiling_status_result',
-			success: true,
-			active: status.active,
-			startedAt: status.startedAt,
-			elapsedMs: status.elapsedMs,
-			categories: status.categories,
-			requestId: message.requestId,
-		});
-	}
-
-	/**
-	 * Handle profiling_stop - stop the capture and write the bundle to `outputPath`.
-	 *
-	 * Unlike the desktop flow (which prompts for a save location), the CLI supplies
-	 * an absolute output path so external scripts can capture -> analyze -> loop
-	 * unattended. The trace is flushed to a temp file first (ending the recording
-	 * overhead immediately), then compressed into the .zip bundle at `outputPath`.
-	 */
-	private async handleProfilingStop(client: WebClient, message: WebClientMessage): Promise<void> {
-		const sendResult = (success: boolean, extra?: Record<string, unknown>) => {
-			this.send(client, {
-				type: 'profiling_stop_result',
-				success,
-				...extra,
-				requestId: message.requestId,
-			});
-		};
-
-		const outputPath = typeof message.outputPath === 'string' ? message.outputPath : '';
-		if (!outputPath) {
-			sendResult(false, { error: 'Missing outputPath' });
-			return;
-		}
-		// The CLI resolves to an absolute path (it knows the user's cwd); the app's
-		// cwd is unrelated, so reject relative paths rather than guess.
-		if (!path.isAbsolute(outputPath)) {
-			sendResult(false, { error: `outputPath must be absolute, got: ${outputPath}` });
-			return;
-		}
-
-		if (!getProfilingStatus().active) {
-			sendResult(false, { error: 'No active profiling recording to stop' });
-			return;
-		}
-
-		const timestamp = new Date().toISOString().replace(/[:.]/g, '-').slice(0, 19);
-		const tracePath = path.join(app.getPath('temp'), `maestro-trace-${timestamp}.json`);
-
-		try {
-			const { durationMs, categories } = await stopProfiling(tracePath);
-			// Ensure the destination directory exists so callers can point at a fresh
-			// output dir per loop iteration without pre-creating it.
-			await fs.mkdir(path.dirname(outputPath), { recursive: true });
-			const finalized = await finalizeCapture(tracePath, outputPath, durationMs, categories);
-			logger.info(`[Profiling] CLI capture saved: ${finalized.path} (${durationMs}ms trace)`);
-			sendResult(true, {
-				path: finalized.path,
-				bundleSizeBytes: finalized.bundleSizeBytes,
-				traceSizeBytes: finalized.traceSizeBytes,
-				durationMs,
-			});
-		} catch (error) {
-			const errMsg = error instanceof Error ? error.message : String(error);
-			sendResult(false, { error: `Failed to save profile: ${errMsg}` });
-		} finally {
-			await fs.rm(tracePath, { force: true }).catch(() => {});
-		}
-	}
-
-	/**
-	 * Handle marketplace_get_manifest - return merged official + local catalog.
-	 */
-	private handleMarketplaceGetManifest(client: WebClient, message: WebClientMessage): void {
-		const refresh = message.refresh === true;
-
-		if (!this.callbacks.getMarketplaceManifest) {
-			this.send(client, {
-				type: 'marketplace_get_manifest_result',
-				success: false,
-				error: 'Marketplace not configured',
-				requestId: message.requestId,
-			});
-			return;
-		}
-
-		this.callbacks
-			.getMarketplaceManifest({ refresh })
-			.then((result) => {
-				if (!result) {
-					this.send(client, {
-						type: 'marketplace_get_manifest_result',
-						success: false,
-						error: 'No manifest available',
-						requestId: message.requestId,
-					});
-					return;
-				}
-				this.send(client, {
-					type: 'marketplace_get_manifest_result',
-					success: true,
-					manifest: result.manifest,
-					fromCache: result.fromCache,
-					cacheAge: result.cacheAge,
-					requestId: message.requestId,
-				});
-			})
-			.catch((error) => {
-				this.reportMarketplaceHandlerError(
-					client,
-					error,
-					'marketplace_get_manifest_result',
-					'marketplace_get_manifest',
-					message,
-					'Failed to load marketplace'
-				);
-			});
-	}
-
-	/**
-	 * Reject a `playbookPath` that points at the local filesystem (absolute
-	 * path or `~`-prefixed) or contains traversal segments / backslashes.
-	 * Web clients can browse the official + local catalog by id, but they
-	 * must never be able to coerce the server into reading arbitrary files
-	 * via the marketplace fetch helpers. Defense-in-depth: downstream
-	 * resolvers also validate, but rejecting at the entry point keeps
-	 * future code changes from re-opening the bypass.
-	 */
-	private isUntrustedLocalPath(playbookPath: string): boolean {
-		if (
-			playbookPath.startsWith('/') ||
-			playbookPath.startsWith('\\') ||
-			playbookPath.startsWith('~/') ||
-			playbookPath.startsWith('~\\') ||
-			/^[a-zA-Z]:[\\/]/.test(playbookPath)
-		) {
-			return true;
-		}
-		// Reject any backslash anywhere - official/local manifest paths use
-		// forward slashes, so a backslash is either a Windows-style absolute
-		// fragment or a deliberate normalization-bypass attempt.
-		if (playbookPath.includes('\\')) {
-			return true;
-		}
-		// Reject `.` / `..` segments. `path.resolve()` collapses these later
-		// but checking up front prevents a relative-traversal payload from
-		// reaching downstream code at all.
-		const segments = playbookPath.split('/');
-		for (const segment of segments) {
-			if (segment === '.' || segment === '..') {
-				return true;
-			}
-		}
-		return false;
-	}
-
-	/**
-	 * Send a typed `marketplace_*_result` failure rather than a generic
-	 * `error` frame. Clients that wait on the typed result type would
-	 * otherwise miss the failure and time out (coderabbit feedback).
-	 */
-	private sendMarketplaceFailure(
-		client: WebClient,
-		type:
-			| 'marketplace_get_manifest_result'
-			| 'marketplace_get_document_result'
-			| 'marketplace_get_readme_result'
-			| 'marketplace_import_playbook_result',
-		error: string,
-		message: WebClientMessage,
-		extra?: Record<string, unknown>
-	): void {
-		this.send(client, {
-			type,
-			success: false,
-			error,
-			requestId: message.requestId,
-			...extra,
-		});
-	}
-
-	/**
-	 * Report an unexpected marketplace-handler exception to Sentry, then
-	 * send a typed failure to the client. Mirrors `reportHandlerError` but
-	 * preserves the `marketplace_*_result` typing the mobile client waits
-	 * on. Without the Sentry capture step, transient production faults in
-	 * the marketplace flow stay invisible because the client only sees the
-	 * typed failure.
-	 */
-	private reportMarketplaceHandlerError(
-		client: WebClient,
-		error: unknown,
-		type:
-			| 'marketplace_get_manifest_result'
-			| 'marketplace_get_document_result'
-			| 'marketplace_get_readme_result'
-			| 'marketplace_import_playbook_result',
-		handler: string,
-		message: WebClientMessage,
-		userMessagePrefix: string,
-		extra?: Record<string, unknown>
-	): void {
-		const err = error instanceof Error ? error : new Error(String(error));
-		captureException(err, { extra: { area: 'web-server', handler, ...extra } });
-		this.sendMarketplaceFailure(
-			client,
-			type,
-			`${userMessagePrefix}: ${err.message}`,
-			message,
-			extra
-		);
-	}
-
-	/**
-	 * Handle marketplace_get_document - fetch a single document's content.
-	 */
-	private handleMarketplaceGetDocument(client: WebClient, message: WebClientMessage): void {
-		const playbookPath = message.playbookPath;
-		const filename = message.filename;
-
-		if (typeof playbookPath !== 'string' || playbookPath.trim() === '') {
-			this.sendMarketplaceFailure(
-				client,
-				'marketplace_get_document_result',
-				'Missing or invalid playbookPath',
-				message
-			);
-			return;
-		}
-		if (this.isUntrustedLocalPath(playbookPath)) {
-			this.sendMarketplaceFailure(
-				client,
-				'marketplace_get_document_result',
-				'Local filesystem paths are not allowed from web clients',
-				message
-			);
-			return;
-		}
-		if (typeof filename !== 'string' || filename.trim() === '') {
-			this.sendMarketplaceFailure(
-				client,
-				'marketplace_get_document_result',
-				'Missing or invalid filename',
-				message
-			);
-			return;
-		}
-		if (filename.includes('..') || filename.includes('/') || filename.includes('\\')) {
-			this.sendMarketplaceFailure(
-				client,
-				'marketplace_get_document_result',
-				'Invalid filename',
-				message
-			);
-			return;
-		}
-
-		if (!this.callbacks.getMarketplaceDocument) {
-			this.sendMarketplaceFailure(
-				client,
-				'marketplace_get_document_result',
-				'Marketplace not configured',
-				message
-			);
-			return;
-		}
-
-		this.callbacks
-			.getMarketplaceDocument(playbookPath, filename)
-			.then((result) => {
-				if (!result) {
-					this.sendMarketplaceFailure(
-						client,
-						'marketplace_get_document_result',
-						'Marketplace not configured',
-						message
-					);
-					return;
-				}
-				this.send(client, {
-					type: 'marketplace_get_document_result',
-					success: true,
-					content: result.content,
-					requestId: message.requestId,
-				});
-			})
-			.catch((error) => {
-				this.reportMarketplaceHandlerError(
-					client,
-					error,
-					'marketplace_get_document_result',
-					'marketplace_get_document',
-					message,
-					'Failed to fetch document',
-					{ playbookPath, filename }
-				);
-			});
-	}
-
-	/**
-	 * Handle marketplace_get_readme - fetch a playbook's README.
-	 */
-	private handleMarketplaceGetReadme(client: WebClient, message: WebClientMessage): void {
-		const playbookPath = message.playbookPath;
-		if (typeof playbookPath !== 'string' || playbookPath.trim() === '') {
-			this.sendMarketplaceFailure(
-				client,
-				'marketplace_get_readme_result',
-				'Missing or invalid playbookPath',
-				message
-			);
-			return;
-		}
-		if (this.isUntrustedLocalPath(playbookPath)) {
-			this.sendMarketplaceFailure(
-				client,
-				'marketplace_get_readme_result',
-				'Local filesystem paths are not allowed from web clients',
-				message
-			);
-			return;
-		}
-
-		if (!this.callbacks.getMarketplaceReadme) {
-			this.sendMarketplaceFailure(
-				client,
-				'marketplace_get_readme_result',
-				'Marketplace not configured',
-				message
-			);
-			return;
-		}
-
-		this.callbacks
-			.getMarketplaceReadme(playbookPath)
-			.then((result) => {
-				if (!result) {
-					this.sendMarketplaceFailure(
-						client,
-						'marketplace_get_readme_result',
-						'Marketplace not configured',
-						message
-					);
-					return;
-				}
-				this.send(client, {
-					type: 'marketplace_get_readme_result',
-					success: true,
-					content: result.content,
-					requestId: message.requestId,
-				});
-			})
-			.catch((error) => {
-				this.reportMarketplaceHandlerError(
-					client,
-					error,
-					'marketplace_get_readme_result',
-					'marketplace_get_readme',
-					message,
-					'Failed to fetch README',
-					{ playbookPath }
-				);
-			});
-	}
-
-	/**
-	 * Handle marketplace_import_playbook - import a playbook into the
-	 * session's Auto Run folder. The server resolves both the folder path
-	 * and SSH config from the session, so the mobile client doesn't need
-	 * to send them - and can't lie about them.
-	 */
-	private handleMarketplaceImportPlaybook(client: WebClient, message: WebClientMessage): void {
-		const sessionId = message.sessionId;
-		const playbookId = message.playbookId;
-		const targetFolderName = message.targetFolderName;
-
-		if (typeof sessionId !== 'string' || sessionId.trim() === '') {
-			this.sendMarketplaceFailure(
-				client,
-				'marketplace_import_playbook_result',
-				'Missing sessionId',
-				message
-			);
-			return;
-		}
-		if (typeof playbookId !== 'string' || playbookId.trim() === '') {
-			this.sendMarketplaceFailure(
-				client,
-				'marketplace_import_playbook_result',
-				'Missing playbookId',
-				message,
-				{ sessionId }
-			);
-			return;
-		}
-		if (typeof targetFolderName !== 'string' || targetFolderName.trim() === '') {
-			this.sendMarketplaceFailure(
-				client,
-				'marketplace_import_playbook_result',
-				'Missing targetFolderName',
-				message,
-				{ sessionId }
-			);
-			return;
-		}
-		// Reject path separators and traversal so the import folder cannot
-		// escape the session's Auto Run root. The service-layer guard
-		// (assertSafeTargetFolderName) is the source of truth, but
-		// short-circuiting here returns a cleaner WebSocket error code.
-		const trimmedFolder = targetFolderName.trim();
-		if (
-			trimmedFolder.includes('..') ||
-			trimmedFolder.includes('/') ||
-			trimmedFolder.includes('\\') ||
-			trimmedFolder.startsWith('~') ||
-			/^[a-zA-Z]:[\\/]/.test(trimmedFolder)
-		) {
-			this.sendMarketplaceFailure(
-				client,
-				'marketplace_import_playbook_result',
-				'targetFolderName must be a single folder name without separators',
-				message,
-				{ sessionId }
-			);
-			return;
-		}
-
-		if (!this.callbacks.importMarketplacePlaybook) {
-			this.sendMarketplaceFailure(
-				client,
-				'marketplace_import_playbook_result',
-				'Marketplace import not configured',
-				message,
-				{ sessionId }
-			);
-			return;
-		}
-
-		this.callbacks
-			.importMarketplacePlaybook(sessionId, playbookId, trimmedFolder)
-			.then((result) => {
-				this.send(client, {
-					type: 'marketplace_import_playbook_result',
-					success: result.success,
-					error: result.error,
-					playbook: result.playbook,
-					importedDocs: result.importedDocs,
-					importedAssets: result.importedAssets,
-					sessionId,
-					requestId: message.requestId,
-				});
-			})
-			.catch((error) => {
-				this.reportMarketplaceHandlerError(
-					client,
-					error,
-					'marketplace_import_playbook_result',
-					'marketplace_import_playbook',
-					message,
-					'Import failed',
-					{ sessionId, playbookId, targetFolderName: trimmedFolder }
-				);
-			});
-	}
-
-	/**
-	 * Handle list_desktop_sessions message - enumerate every open AI tab across
-	 * desktop agents. Stateless read backed by the persisted session store; no
-	 * subscription side-effects so external pollers (Maestro-Discord, Cue) can
-	 * call this every few seconds without leaking state into the desktop.
-	 */
-	private handleListDesktopSessions(client: WebClient, message: WebClientMessage): void {
-		const sessions = this.callbacks.listDesktopSessions ? this.callbacks.listDesktopSessions() : [];
-		this.send(client, {
-			type: 'desktop_sessions_list',
-			success: true,
-			sessions,
-			requestId: message.requestId,
-		});
-	}
-
-	/**
-	 * Handle plugins_list_tools - project the registered plugin `tools`
-	 * contributions into MCP tool defs for the `maestro-cli mcp serve` bridge.
-	 * Returns an MCP-safe `name` (namespaced id, `/`->`__`) plus the real
-	 * `toolId` so the bridge can reverse-map on call. Empty when the plugins flag
-	 * is off or no manager is wired.
-	 */
-	private handlePluginsListTools(client: WebClient, message: WebClientMessage): void {
-		const manager = getActivePluginManager();
-		let tools: Array<{
-			name: string;
-			toolId: string;
-			description: string;
-			inputSchema: Record<string, unknown>;
-		}> = [];
-		if (manager && isPluginsFeatureEnabled()) {
-			try {
-				tools = manager.getContributions().tools.map((t) => ({
-					name: t.id.replace(/\//g, '__').replace(/[^a-zA-Z0-9_-]/g, '_'),
-					toolId: t.id,
-					description: t.description,
-					inputSchema:
-						t.inputSchema && typeof t.inputSchema === 'object' && !Array.isArray(t.inputSchema)
-							? t.inputSchema
-							: { type: 'object' },
-				}));
-			} catch (error) {
-				const reason = error instanceof Error ? error.message : String(error);
-				logger.warn(`[Web] plugins_list_tools failed: ${reason}`, LOG_CONTEXT);
-			}
-		}
-		this.send(client, {
-			type: 'plugins_list_tools_result',
-			success: true,
-			tools,
-			requestId: message.requestId,
-		});
-	}
-
-	/**
-	 * Handle plugins_call_tool - risk-gate a model-initiated plugin tool call,
-	 * then invoke it via the broker. The toolId MUST be a declared `tools`
-	 * contribution (never an arbitrary command handler), and risk is rated on the
-	 * model's ARGUMENTS via the shared Pianola gate - a HIGH verdict is surfaced
-	 * and NEVER executed. Tool failures: `{ ok:false, error }`; blocks: `{ blocked:true }`.
-	 */
-	private async handlePluginsCallTool(client: WebClient, message: WebClientMessage): Promise<void> {
-		const respond = (extra: Record<string, unknown>): void =>
-			this.send(client, {
-				type: 'plugins_call_tool_result',
-				requestId: message.requestId,
-				...extra,
-			});
-		const toolId = typeof message.toolId === 'string' ? message.toolId : '';
-		if (!toolId) {
-			respond({ ok: false, error: 'Missing toolId' });
-			return;
-		}
-		const manager = getActivePluginManager();
-		if (!manager || !isPluginsFeatureEnabled()) {
-			respond({ ok: false, error: 'PluginsDisabled' });
-			return;
-		}
-		const declaredTool = manager.getContributions().tools.find((t) => t.id === toolId);
-		if (!declaredTool) {
-			// Only DECLARED `tools` are model-callable; never let a tools/call name
-			// resolve to an arbitrary command handler in the sandbox's shared map.
-			respond({ ok: false, error: `Unknown tool: ${toolId}` });
-			return;
-		}
-		const args = 'args' in message ? message.args : undefined;
-		let argText = '';
-		try {
-			argText = JSON.stringify(args ?? {});
-		} catch {
-			argText = '';
-		}
-		// Rate risk on the declared tool's human name + description + the model's
-		// args: catches a destructive tool by identity AND destructive args, without
-		// the slug noise of the raw toolId. Follow-up: per-tool risk metadata + a
-		// user-approval path for HIGH instead of a hard block.
-		const riskText = `${declaredTool.name} ${declaredTool.description} ${argText}`;
-		const verdict = evaluatePluginDispatch(riskText);
-		if (!verdict.eligible) {
-			respond({ ok: false, blocked: true, risk: verdict.risk, reason: verdict.reason });
-			return;
-		}
-		try {
-			const result = await manager.invokeTool(toolId, args);
-			respond({ ok: true, result });
-		} catch (error) {
-			respond({ ok: false, error: error instanceof Error ? error.message : String(error) });
-		}
-	}
-
-	/**
-	 * Handle get_session_history message - return the conversation log for a
-	 * tab, optionally filtered by `sinceMs` (poll cursor) and/or `tail` (cap).
-	 * Errors are returned in the same response type rather than as a generic
-	 * `error` so the CLI's request/response pairing stays deterministic.
-	 */
-	private handleGetSessionHistory(client: WebClient, message: WebClientMessage): void {
-		const tabId = typeof message.tabId === 'string' ? message.tabId : undefined;
-		if (!tabId) {
-			this.send(client, {
-				type: 'session_history_result',
-				success: false,
-				error: 'Missing tabId',
-				code: 'MISSING_TAB_ID',
-				requestId: message.requestId,
-			});
-			return;
-		}
-
-		if (!this.callbacks.getSessionHistory) {
-			this.send(client, {
-				type: 'session_history_result',
-				success: false,
-				error: 'Session history not configured',
-				code: 'NOT_CONFIGURED',
-				requestId: message.requestId,
-			});
-			return;
-		}
-
-		const sinceMs =
-			typeof message.sinceMs === 'number' && Number.isFinite(message.sinceMs)
-				? message.sinceMs
-				: undefined;
-		const tail =
-			typeof message.tail === 'number' && Number.isFinite(message.tail) && message.tail >= 0
-				? Math.floor(message.tail)
-				: undefined;
-
-		const result = this.callbacks.getSessionHistory(tabId, { sinceMs, tail });
-		if (!result) {
-			this.send(client, {
-				type: 'session_history_result',
-				success: false,
-				error: `Tab not found: ${tabId}`,
-				code: 'TAB_NOT_FOUND',
-				requestId: message.requestId,
-			});
-			return;
-		}
-
-		this.send(client, {
-			type: 'session_history_result',
-			success: true,
-			tabId: result.tabId,
-			sessionId: result.sessionId,
-			agentId: result.agentId,
-			agentSessionId: result.agentSessionId,
-			projectPath: result.projectPath,
-			messages: result.messages,
-			requestId: message.requestId,
 		});
 	}
